### 2025-11-03 #8971 @mbobra
@crvernon I do not know R or C++, so I won't be too valuable here. Please do assign me to an incoming submission in Python!
