### 2025-11-12 4c2d183
Add CONTRIBUTING.md and CODE_OF_CONDUCT.md for JOSS submission
