+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,45 @@
+# Code of Conduct
+
+## Our Pledge
+
+We aim to maintain an open, friendly, and professional environment within the **corrselect** project.  
+Everyone taking part—contributors, maintainers, and users—should feel welcome and respected, regardless of background, identity, or experience level.
+
+## Our Standards
+
+Examples of behavior that helps create a good environment:
+
+- Communicating clearly and respectfully  
+- Being open to different opinions and approaches  
+- Offering constructive feedback  
+- Helping others learn and contribute  
+- Staying focused on collaboration and shared goals  
+
+Examples of behavior that is not acceptable:
+
+- Personal attacks or insulting language  
+- Disrespectful or dismissive comments  
+- Sharing private information without consent  
+- Any form of harassment or hostility toward others  
+
+## Responsibilities
+
+Project maintainers are responsible for clarifying standards of acceptable behavior and taking fair action when necessary.  
+They may edit, remove, or reject contributions that violate this Code of Conduct or disrupt collaboration.
+
+## Scope
+
+This Code of Conduct applies to all project spaces, including discussions, issues, pull requests, and any community interactions related to corrselect.
+
+## Enforcement
+
+If you experience or witness behavior that violates this Code of Conduct, please contact the maintainer privately.  
+Reports will be handled with discretion, and maintainers will take appropriate action as needed.
+
+## Attribution
+
+This Code of Conduct is adapted from the [Contributor Covenant](https://www.contributor-covenant.org), version 2.1, and simplified to better fit the tone of an open scientific software project.
+
+---
+
+**Contact:** gilles.colling@univie.ac.at
