+++ CONTRIBUTING.md
@@ -1,0 +1,245 @@
+# Contribution Guidelines
+
+First of all, thank you very much for taking the time to contribute
+to the **corrselect** project!
+
+This document provides guidelines for contributing to corrselect—its codebase and documentation.  
+These guidelines are meant to guide you, not to restrict you.  
+If in doubt, use your best judgment and feel free to propose improvements through an issue or pull request.
+
+#### Table Of Contents
+
+- [Code of Conduct](#code-of-conduct)
+- [Installation](#installation)
+  - [Obtaining the source](#obtaining-the-source)
+  - [Setting up your R environment](#setting-up-your-r-environment)
+  - [Installing from source](#installing-from-source)
+- [Testing](#testing)
+- [Documentation](#documentation)
+  - [Install dependencies](#install-dependencies)
+  - [Building the documentation](#building-the-documentation)
+  - [Design of the docs](#design-of-the-docs)
+- [Project organization](#project-organization)
+- [Contributing workflow](#contributing-workflow)
+- [Style guidelines](#style-guidelines)
+- [Pull request checklist](#pull-request-checklist)
+- [Reporting bugs](#reporting-bugs)
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by our **Code of Conduct** (`CODE_OF_CONDUCT.md`).  
+By participating, you are expected to uphold this code and maintain a respectful, inclusive environment.
+
+## Installation
+
+This installation guide is focused on development.  
+For regular installation, please see the [README](./README.md).
+
+### Obtaining the source
+
+Clone the corrselect repository:
+
+```bash
+git clone https://github.com/gcol33/corrselect.git
+cd corrselect
+```
+
+If you work on a development branch:
+
+```bash
+git checkout dev
+git pull origin dev
+```
+
+### Setting up your R environment
+
+corrselect is an R package that uses C++ code via Rcpp.
+
+1. **Install required tools**
+   - R (≥ 4.0)
+   - Rtools (Windows) or Xcode Command Line Tools (macOS)
+   - Git
+   - An editor or IDE (RStudio, VS Code, etc.)
+
+2. **Install development dependencies**
+   ```r
+   install.packages(c("devtools", "roxygen2", "testthat", "rmarkdown", "knitr", "pkgdown"))
+   ```
+
+3. **Load the development build**
+   ```r
+   devtools::load_all()
+   ```
+
+### Installing from source
+
+Build and install the package locally:
+
+```r
+devtools::install()
+```
+
+If you modify C++ code, rebuild the DLL before reinstalling:
+
+```r
+devtools::clean_dll()
+devtools::install()
+```
+
+Regenerate documentation with:
+
+```r
+devtools::document()
+```
+
+## Testing
+
+corrselect uses **testthat** for testing.  
+All tests are located in `tests/testthat/`.
+
+Run the full test suite:
+
+```r
+devtools::test()
+```
+
+Run a complete package check:
+
+```r
+devtools::check()
+```
+
+Run a subset of tests during development:
+
+```r
+testthat::test_dir("tests/testthat")
+testthat::test_file("tests/testthat/test-corrSelect.R")
+```
+
+Guidelines:
+- Keep tests fast and reproducible.
+- Use `set.seed()` for random data.
+- Include edge cases and expected failures.
+- Prefer small examples to large datasets.
+
+## Documentation
+
+### Install dependencies
+
+```r
+install.packages(c("rmarkdown", "knitr", "pkgdown"))
+```
+
+### Building the documentation
+
+Build vignettes:
+
+```r
+devtools::build_vignettes()
+```
+
+Build the pkgdown site locally:
+
+```r
+pkgdown::build_site()
+```
+
+The generated site is saved in the `docs/` directory.  
[…]
