### 2025-11-12 #8971 @gcol33
Thanks, @Nikoleta-v3.

I’ve added CONTRIBUTING.md and CODE_OF_CONDUCT.md to the repository root.
The CONTRIBUTING file includes details on running tests (devtools::test(), devtools::check()), building vignettes, and generating documentation with pkgdown.

For reviewers, here are five suggestions that align well with corrselect’s focus on R-based statistical methods and ecological applications:

@pederengelstad

@mhesselbarth

@ahasverus

@dbarneche

@trashbirdecology

Let me know if anything else is needed!
