### 2025-11-12 b71968a
Refactor wording for clarity in CODE_OF_CONDUCT.md
