### 2025-11-13 #8971 @Nikoleta-v3
Amazing! Thank you, @gcol33 ✨  the contributing file and code of conduct look great 😄
