### 2025-11-20 1265a8f
Fix modelPrune infinite loop and eliminate all test warnings (CRAN-ready)
## Bug Fixes

- **modelPrune**: Fixed infinite loop when VIF computation encountered perfect multicollinearity
  - Added proper handling of Inf and NA VIF values in pruning loop
  - Clamped extreme R² values (> 0.9999) to prevent division by near-zero
  - Added safety checks to prevent removing all variables

- **modelPrune**: Fixed design matrix extraction for lme4 and glmmTMB engines
  - Now uses stats::model.matrix() for all engines (more robust)
  - Eliminated "Could not find columns" warnings

- **Test suite**: All 
[…]
