+++ R/RcppExports.R
@@ -13,6 +13,10 @@ runELS <- function(corMatrix, threshold, forcedVec) {
     .Call(`_corrselect_runELS`, corMatrix, threshold, forcedVec)
 }
 
+greedyPruneBackend <- function(assoc_matrix, threshold, force_in = NULL) {
+    .Call(`_corrselect_greedyPruneBackend`, assoc_matrix, threshold, force_in)
+}
+
 runSpectral <- function(corMatrix, threshold, forcedVec, lengthLimit, k_param = NULL) {
     .Call(`_corrselect_runSpectral`, corMatrix, threshold, forcedVec, lengthLimit, k_param)
 }
