+++ R/data.R
@@ -1,0 +1,179 @@
+#' Example Bioclimatic Data for Ecological Modeling
+#'
+#' A simulated dataset representing 50 WorldClim-like bioclimatic variables
+#' measured at 100 geographic locations, with species richness as the response variable.
+#' Variables are organized into correlated blocks representing temperature,
+#' precipitation, seasonality, and elevation.
+#'
+#' @format A data frame with 100 rows and 51 variables:
+#' \describe{
+#'   \item{species_richness}{Integer. Number of species observed (response variable)}
+#'   \item{bio1_temp through bio15_temp}{Numeric. Temperature-related variables (highly correlated, r ≈ 0.85)}
+#'   \item{bio16_precip through bio30_precip}{Numeric. Precipitation-related variables (moderately correlated, r ≈ 0.70)}
+#'   \item{bio31_season through bio45_season}{Numeric. Seasonality variables (weakly correlated, r ≈ 0.50)}
+#'   \item{bio46_elev through bio50_elev}{Numeric. Elevation variables (independent, r ≈ 0)}
+#' }
+#'
+#' @details
+#' This dataset demonstrates a common problem in ecological modeling: many bioclimatic
+#' predictors are highly correlated within groups (e.g., different temperature metrics),
+#' leading to multicollinearity issues. The species richness response depends on a subset
+#' of predictors from each group.
+#'
+#' **Use case**: Demonstrating `corrPrune()` for reducing correlated environmental
+#' predictors before fitting species distribution models.
+#'
+#' @source Simulated data based on typical WorldClim variable structures
+#'
+#' @seealso [corrPrune()], [modelPrune()]
+#'
+#' @examples
+#' data(bioclim_example)
+#'
+#' # Remove highly correlated bioclimatic variables
+#' pruned <- corrPrune(bioclim_example[, -1], threshold = 0.7)
+#' ncol(pruned)  # Reduced from 50 to ~15 variables
+#'
+#' # Model-based pruning with VIF
+#' model_data <- modelPrune(species_richness ~ .,
+#'                          data = bioclim_example,
+#'                          limit = 5)
+#' attr(model_data, "selected_vars")
+"bioclim_example"
+
+#' Example Survey Data for Social Science Research
+#'
+#' A simulated questionnaire dataset with 30 Likert-scale items measuring three
+#' latent constructs (satisfaction, engagement, loyalty), plus demographic variables
+#' and an overall satisfaction score.
+#'
+#' @format A data frame with 200 rows and 35 variables:
+#' \describe{
+#'   \item{respondent_id}{Integer. Unique respondent identifier}
+#'   \item{age}{Integer. Respondent age (18-75 years)}
+#'   \item{gender}{Factor. Gender (Male, Female, Other)}
+#'   \item{education}{Ordered factor. Education level (High School, Bachelor, Master, PhD)}
+#'   \item{overall_satisfaction}{Integer. Overall satisfaction score (0-100)}
+#'   \item{satisfaction_1 through satisfaction_10}{Ordered factor. Satisfaction items (1-7 Likert scale)}
+#'   \item{engagement_1 through engagement_10}{Ordered factor. Engagement items (1-7 Likert scale)}
+#'   \item{loyalty_1 through loyalty_10}{Ordered factor. Loyalty items (1-7 Likert scale)}
+#' }
+#'
+#' @details
+#' This dataset represents a common scenario in survey research: multiple items
+#' measuring similar constructs lead to redundancy and multicollinearity. Items
+#' within each construct are correlated (satisfaction, engagement, loyalty), and
[…]
