+++ R/modelPrune.R
@@ -1,0 +1,708 @@
+#' Model-Based Predictor Pruning
+#'
+#' `modelPrune()` performs iterative removal of fixed-effect predictors based on
+#' model diagnostics (e.g., VIF) until all remaining predictors satisfy a
+#' specified threshold. It supports linear models, generalized linear models,
+#' and mixed models.
+#'
+#' @param formula A model formula specifying the response and predictors.
+#'   May include random effects for mixed models (e.g., `y ~ x1 + x2 + (1|group)`).
+#' @param data A data.frame containing the variables in the formula.
+#' @param engine Either a character string for built-in engines, or a list defining a custom engine.
+#'
+#'   **Built-in engines** (character string):
+#'   - `"lm"` (default): Linear models via `stats::lm()`
+#'   - `"glm"`: Generalized linear models via `stats::glm()` (requires `family` argument)
+#'   - `"lme4"`: Mixed models via `lme4::lmer()` or `lme4::glmer()` (requires lme4 package)
+#'   - `"glmmTMB"`: Generalized linear mixed models via `glmmTMB::glmmTMB()` (requires glmmTMB package)
+#'
+#'   **Custom engine** (named list with required components):
+#'   - `fit`: function(formula, data, ...) that returns a fitted model object
+#'   - `diagnostics`: function(model, fixed_effects) that returns a named numeric vector
+#'     of diagnostic scores (one per fixed effect, higher values = worse)
+#'   - `name` (optional): character string used in error messages (default: "custom")
+#' @param criterion Character string specifying the diagnostic criterion for pruning.
+#'   For built-in engines, only `"vif"` (Variance Inflation Factor) is supported.
+#'   For custom engines, this parameter is ignored (diagnostics are computed by the
+#'   engine's `diagnostics` function). Default: `"vif"`.
+#' @param limit Numeric scalar. Maximum allowed value for the criterion.
+#'   Predictors with diagnostic values exceeding this limit are iteratively removed.
+#'   Default: 5 (common VIF threshold).
+#' @param force_in Character vector of predictor names that must be retained
+#'   in the final model. These variables will not be removed during pruning.
+#'   Default: NULL.
+#' @param max_steps Integer. Maximum number of pruning iterations. If NULL (default),
+#'   pruning continues until all diagnostics are below the limit or no more
+#'   removable predictors remain.
+#' @param ... Additional arguments passed to the modeling function (e.g., `family`
+#'   for glm/glmer, control parameters for lme4/glmmTMB).
+#'
+#' @return A data.frame containing only the retained predictors (and response).
+#'   The result has the following attributes:
+#'   \describe{
+#'     \item{selected_vars}{Character vector of retained predictor names}
+#'     \item{removed_vars}{Character vector of removed predictor names (in order of removal)}
+#'     \item{engine}{Character string indicating which engine was used}
+#'     \item{criterion}{Character string indicating which criterion was used}
+#'     \item{limit}{The threshold value used}
+#'     \item{final_model}{The final fitted model object (optional)}
+#'   }
+#'
+#' @details
+#' `modelPrune()` works by:
+#' 1. Parsing the formula to identify fixed-effect predictors
+#' 2. Fitting the initial model
+#' 3. Computing diagnostics for each fixed-effect predictor
+#' 4. Checking feasibility of `force_in` constraints
+#' 5. Iteratively removing the predictor with the worst diagnostic value
+#'    (excluding `force_in` variables) until all diagnostics ≤ `limit`
+#' 6. Returning the pruned data frame
+#'
+#' **Random Effects**: For mixed models (lme4, glmmTMB), only fixed-effect
[…]
