### 2025-11-24 6b64043
Update cran-comments.md for version 3.0.0 submission (not a new package)
