### 2025-11-24 9d91b5b
Clean repository: remove build artifacts and old CRAN submission file
- Remove corrselect.Rcheck/ directory (should never be committed)
- Remove corrselect_2.0.1.tar.gz (old package tarball)
- Remove CRAN-SUBMISSION (outdated, will be regenerated)
- Update .Rbuildignore to exclude future tarballs
