+++ corrselect.Rcheck/00_pkg_src/corrselect/R/CorrCombo.R
@@ -1,256 +1,0 @@
-#' @title CorrCombo S4 class
-#'
-#' @description
-#' Holds the result of \code{\link{corrSelect}} or \code{\link{MatSelect}}: a list of valid variable combinations
-#' and their correlation statistics.
-#'
-#' This class stores all subsets of variables that meet the specified correlation constraint,
-#' along with metadata such as the algorithm used, correlation method(s), variables forced into every subset,
-#' and summary statistics for each combination.
-#'
-#' @name CorrCombo
-#' @docType class
-#' @aliases CorrCombo-class
-#'
-#' @slot subset_list   A list of character vectors. Each vector is a valid subset (variable names).
-#' @slot avg_corr      A numeric vector. Average absolute correlation within each subset.
-#' @slot min_corr      A numeric vector. Minimum pairwise absolute correlation in each subset.
-#' @slot max_corr      A numeric vector. Maximum pairwise absolute correlation within each subset.
-#' @slot names         Character vector of all variable names used for decoding.
-#' @slot threshold     Numeric scalar. The correlation threshold used during selection.
-#' @slot forced_in     Character vector. Variable names that were forced into each subset.
-#' @slot search_type   Character string. One of \code{"els"} or \code{"bron-kerbosch"}.
-#' @slot cor_method    Character string. Either a single method (e.g. "pearson") or "mixed" if multiple methods used.
-#' @slot n_rows_used   Integer. Number of rows used for computing the correlation matrix (after removing missing values).
-#'
-#' @seealso \code{\link{corrSelect}}, \code{\link{MatSelect}}, \code{\link{corrSubset}}
-#'
-#' @examples
-#' show(new("CorrCombo",
-#'   subset_list = list(c("A", "B"), c("A", "C")),
-#'   avg_corr = c(0.2, 0.3),
-#'   min_corr = c(0.1, 0.2),
-#'   max_corr = c(0.3, 0.4),
-#'   names = c("A", "B", "C"),
-#'   threshold = 0.5,
-#'   forced_in = character(),
-#'   search_type = "els",
-#'   cor_method = "mixed",
-#'   n_rows_used = as.integer(5)
-#' ))
-#'
-#' @importFrom methods new
-#' @export
-setClass(
-  "CorrCombo",
-  slots = list(
-    subset_list   = "list",
-    avg_corr      = "numeric",
-    min_corr      = "numeric",
-    max_corr      = "numeric",
-    names         = "character",
-    threshold     = "numeric",
-    forced_in     = "character",
-    search_type   = "character",
-    cor_method    = "character",
-    n_rows_used   = "integer"
-  ),
-  validity = function(object) {
-    n <- length(object@subset_list)
-    check_lengths <- function(slot_value) {
-      len <- length(slot_value)
-      len == 0 || len == n
-    }
-    if (!check_lengths(object@avg_corr)) return("avg_corr must match subset_list length or be empty.")
-    if (!check_lengths(object@min_corr)) return("min_corr must match subset_list length or be empty.")
-    if (!check_lengths(object@max_corr)) return("max_corr must match subset_list length or be empty.")
-    TRUE
-  }
-)
-
-#' @rdname CorrCombo
-#' @aliases show,CorrCombo-method
-#' @importMethodsFrom methods show
-#' @exportMethod show
-#' @param object A \code{CorrCombo} object to be printed.
-#' @title CorrCombo Class
-#'
-#' @description
-#' An S4 class that stores the result of correlation-based subset selection.
-#'
[…]
