### 2025-11-24 c5df70d
Fix survey_example documentation and add MASS/caret to Suggests
- Fix survey_example example to use only numeric columns with corrPrune()
- Add MASS and caret to Suggests (used in vignettes)
- Regenerate documentation
