### 2025-11-24 492ad5e
Fix data documentation: list all column names explicitly
- Replace 'through' shorthand with explicit comma-separated lists
- Eliminates code/documentation mismatch WARNING in R CMD check
- All four datasets updated: bioclim_example, survey_example, genes_example, longitudinal_example
