### 2025-11-24 a5eee59
Update to v3.0.0: 19 WorldClim variables + enhanced vignette graphics
- Change bioclim_example dataset from 50 to 19 official WorldClim bioclimatic variables
- Update all documentation with explicit BIO1-BIO19 descriptions
- Fix survey_example documentation to use assocSelect()
- Add 5 new visualizations to workflows vignette:
  * Correlation distribution histogram with threshold line
  * Before/after correlation histograms showing pruning effect
  * Model comparison barplots (predictors & Adj R²)
  * Survey construct coverage barplots
  * Gene dimensionality reduction barplot
- Update all vig
[…]
