+++ R/data.R
@@ -1,38 +1,57 @@
 #' Example Bioclimatic Data for Ecological Modeling
 #'
-#' A simulated dataset representing 50 WorldClim-like bioclimatic variables
-#' measured at 100 geographic locations, with species richness as the response variable.
-#' Variables are organized into correlated blocks representing temperature,
-#' precipitation, seasonality, and elevation.
+#' A simulated dataset with the 19 WorldClim bioclimatic variables
+#' (https://www.worldclim.org/data/bioclim.html) measured at 100 geographic locations,
+#' with species richness as the response variable. Variables are organized into
+#' correlated blocks representing temperature (BIO1-BIO11) and precipitation (BIO12-BIO19).
 #'
-#' @format A data frame with 100 rows and 51 variables:
+#' @format A data frame with 100 rows and 20 variables:
 #' \describe{
 #'   \item{species_richness}{Integer. Number of species observed (response variable)}
-#'   \item{bio1_temp, bio2_temp, bio3_temp, bio4_temp, bio5_temp, bio6_temp, bio7_temp, bio8_temp, bio9_temp, bio10_temp, bio11_temp, bio12_temp, bio13_temp, bio14_temp, bio15_temp}{Numeric. Temperature-related variables (highly correlated, r ≈ 0.85)}
-#'   \item{bio16_precip, bio17_precip, bio18_precip, bio19_precip, bio20_precip, bio21_precip, bio22_precip, bio23_precip, bio24_precip, bio25_precip, bio26_precip, bio27_precip, bio28_precip, bio29_precip, bio30_precip}{Numeric. Precipitation-related variables (moderately correlated, r ≈ 0.70)}
-#'   \item{bio31_season, bio32_season, bio33_season, bio34_season, bio35_season, bio36_season, bio37_season, bio38_season, bio39_season, bio40_season, bio41_season, bio42_season, bio43_season, bio44_season, bio45_season}{Numeric. Seasonality variables (weakly correlated, r ≈ 0.50)}
-#'   \item{bio46_elev, bio47_elev, bio48_elev, bio49_elev, bio50_elev}{Numeric. Elevation variables (independent, r ≈ 0)}
+#'   \item{BIO1}{Numeric. Annual Mean Temperature}
+#'   \item{BIO2}{Numeric. Mean Diurnal Range}
+#'   \item{BIO3}{Numeric. Isothermality}
+#'   \item{BIO4}{Numeric. Temperature Seasonality}
+#'   \item{BIO5}{Numeric. Max Temperature of Warmest Month}
+#'   \item{BIO6}{Numeric. Min Temperature of Coldest Month}
+#'   \item{BIO7}{Numeric. Temperature Annual Range}
+#'   \item{BIO8}{Numeric. Mean Temperature of Wettest Quarter}
+#'   \item{BIO9}{Numeric. Mean Temperature of Driest Quarter}
+#'   \item{BIO10}{Numeric. Mean Temperature of Warmest Quarter}
+#'   \item{BIO11}{Numeric. Mean Temperature of Coldest Quarter}
+#'   \item{BIO12}{Numeric. Annual Precipitation}
+#'   \item{BIO13}{Numeric. Precipitation of Wettest Month}
+#'   \item{BIO14}{Numeric. Precipitation of Driest Month}
+#'   \item{BIO15}{Numeric. Precipitation Seasonality}
+#'   \item{BIO16}{Numeric. Precipitation of Wettest Quarter}
+#'   \item{BIO17}{Numeric. Precipitation of Driest Quarter}
+#'   \item{BIO18}{Numeric. Precipitation of Warmest Quarter}
+#'   \item{BIO19}{Numeric. Precipitation of Coldest Quarter}
 #' }
 #'
 #' @details
[…]
