### 2025-11-24 ab4a7a0
Fix vignette formatting: remove emoji bullets and fix VIF plot axis labels
- Replace all emoji/special character bullets (✅, ❌, ⚠️) with standard markdown
- Fix VIF barplot to display all variable names on x-axis
- Increase bottom margin and explicitly set names.arg for proper label display
- Ensure professional, CRAN-ready formatting throughout all vignettes
