### 2025-11-24 6ba8f62
Improve vignette list formatting and clarify combined workflow guidance
- Fix list formatting: add blank lines and use colons for better readability
- Clarify when to use corrPrune + Boruta workflow (p > 50)
- Add concrete examples of dimensionality reduction (e.g., 100 → 40 variables)
- Add note that smaller datasets (p < 30) may not need both steps

This improves readability and provides clearer guidance on workflow selection.
