+++ vignettes/comparison.Rmd
@@ -1,560 +1,0 @@
----
-title: "Comparison with Alternatives"
-author: "Gilles Colling"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Comparison with Alternatives}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.width = 7,
-  fig.height = 5
-)
-```
-
-## Overview
-
-This vignette compares corrselect with popular alternatives for variable selection and multicollinearity handling:
-
-1. **Manual correlation filtering**
-2. **caret::findCorrelation()**
-3. **Boruta (feature selection)**
-4. **glmnet (regularization)**
-
-**Goal**: Help you choose the right tool for your problem.
-
-**Estimated time**: 10 minutes
-
----
-
-# When to Use corrselect
-
-corrselect is designed for situations where you want to:
-
-- **Remove redundant predictors** based on correlation structure
-- **Reduce multicollinearity** in regression models
-- **Maintain interpretability** (keep predictors, not transform them)
-- **Ensure reproducibility** (deterministic results)
-- **Control which variables to keep** (via `force_in`)
-
----
-
-# Comparison 1: Manual Correlation Filtering
-
-## The Manual Approach
-
-Many analysts use an ad-hoc workflow:
-
-```{r}
-library(corrselect)
-data(mtcars)
-
-# Manual approach
-cor_matrix <- cor(mtcars)
-print(cor_matrix[1:5, 1:5])
-
-# Identify high correlations (> 0.7)
-high_cor <- which(abs(cor_matrix) > 0.7 & abs(cor_matrix) < 1, arr.ind = TRUE)
-head(high_cor)
-
-# Now what? Manually decide which to remove...
-# This is tedious, subjective, and non-reproducible!
-```
-
-**Problems**:
-- Manual decision-making (which variable to drop?)
-- Time-consuming for many variables
-- Not reproducible (different analysts make different choices)
-- No guarantee of optimality
-- Hard to document decision process
-
-## The corrselect Approach
-
-```{r}
-# Automated, reproducible, optimal
-pruned <- corrPrune(mtcars, threshold = 0.7)
-
-# Clear documentation of what was removed
-attr(pruned, "removed_vars")
-attr(pruned, "selected_vars")
-```
-
-**Advantages**:
-- **Fully automated** - no manual decisions
-- **Reproducible** - same result every time
-- **Optimal** - maximizes retained variables (exact mode)
-- **Documented** - clear record of removals
-- **Fast** - greedy mode for large datasets
-
----
-
-# Comparison 2: caret::findCorrelation()
-
-caret provides `findCorrelation()` for removing correlated predictors. Let's compare.
-
-## Side-by-Side Comparison
-
-```{r}
-# Prepare data (numeric only for caret)
-mtcars_num <- mtcars
-
-# caret approach
-if (requireNamespace("caret", quietly = TRUE)) {
-  cor_matrix <- cor(mtcars_num)
-  to_remove_caret <- caret::findCorrelation(cor_matrix, cutoff = 0.7)
-  mtcars_caret <- mtcars_num[, -to_remove_caret]
-
-  cat("caret kept:", ncol(mtcars_caret), "variables\n")
-  cat("caret removed:", colnames(mtcars_num)[to_remove_caret], "\n\n")
-}
-
-# corrselect approach
-mtcars_corrselect <- corrPrune(mtcars_num, threshold = 0.7)
-
[…]
