### 2025-11-25 51559d2
Complete Sprint 4: Publication polish + bullet formatting fixes
Sprint 4 Documentation Improvements:
- Added citation sections with BibTeX to all 5 vignettes
- Added sessionInfo() blocks for reproducibility
- Enhanced quickstart preamble with learning objectives
- Documented corrPrune() selection logic (lowest avg correlation)

Bullet Formatting Fixes:
- Fixed 14 bullet list formatting issues across vignettes
- Added blank lines before bullet lists per Markdown spec
- Created check_bullets.R script for automated validation

New Content:
- Added cor_example dataset with block correlation struct
[…]
