### 2025-11-25 79f570c
Add precomputation infrastructure for vignette optimization
- Created inst/extdata/ directory for precomputed results
- Added precompute_vignette_results.R script
- Generated workflow_ecology.rds (first precomputed dataset)

This infrastructure will speed up future pkgdown builds by
precomputing expensive workflow examples.

Next steps: Complete remaining workflow precomputations and
update vignettes to load precomputed results.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
