+++ data-raw/precompute_vignette_results.R
@@ -1,0 +1,173 @@
+# Script to precompute computationally intensive vignette results
+# Run this script when updating vignettes that have expensive computations
+
+# Load the package in development mode
+devtools::load_all(".")
+
+# Create output directory
+dir.create("inst/extdata", showWarnings = FALSE, recursive = TRUE)
+
+# ============================================================================
+# Workflow 1: Ecology (bioclim_example)
+# ============================================================================
+
+data(bioclim_example)
+
+# Step 1: corrPrune
+bio_clean <- corrPrune(bioclim_example[, -1], threshold = 0.7)
+
+# Step 2: modelPrune with VIF
+bio_model_data <- modelPrune(
+  formula = species_richness ~ .,
+  data = bioclim_example,
+  limit = 5
+)
+
+# Save results
+saveRDS(list(
+  bio_clean = bio_clean,
+  bio_model_data = bio_model_data,
+  original_dim = dim(bioclim_example)
+), file = "inst/extdata/workflow_ecology.rds")
+
+cat("✓ Saved workflow_ecology.rds\n")
+
+# ============================================================================
+# Workflow 2: Survey Data (survey_example)
+# ============================================================================
+
+data(survey_example)
+
+# Exclude ID and response
+survey_predictors <- survey_example[, !(names(survey_example) %in%
+                                         c("respondent_id", "overall_satisfaction"))]
+
+# Prune with force_in (only use numeric columns)
+numeric_cols <- sapply(survey_predictors, is.numeric)
+survey_numeric <- survey_predictors[, numeric_cols, drop = FALSE]
+survey_clean <- corrPrune(
+  survey_numeric,
+  threshold = 0.6,
+  force_in = if("age" %in% names(survey_numeric)) "age" else NULL
+)
+
+# Save results
+saveRDS(list(
+  survey_clean = survey_clean,
+  original_dim = dim(survey_predictors)
+), file = "inst/extdata/workflow_survey.rds")
+
+cat("✓ Saved workflow_survey.rds\n")
+
+# ============================================================================
+# Workflow 3: High-Dimensional Data (genes_example)
+# ============================================================================
+
+data(genes_example)
+
+# Separate response
+genes_predictors <- genes_example[, !names(genes_example) %in% "disease_status"]
+
+# Prune
+genes_clean <- corrPrune(genes_predictors, threshold = 0.8)
+
+# Save results
+saveRDS(list(
+  genes_clean = genes_clean,
+  original_dim = dim(genes_predictors),
+  response = genes_example$disease_status
+), file = "inst/extdata/workflow_genes.rds")
+
+cat("✓ Saved workflow_genes.rds\n")
+
+# ============================================================================
+# Workflow 4: Mixed Models (longitudinal_example)
+# ============================================================================
+
+data(longitudinal_example)
+
+# Prepare data
+longitudinal_predictors <- longitudinal_example[, !names(longitudinal_example) %in%
+                                                  c("subject_id", "timepoint", "outcome")]
+
+# Prune
+longitudinal_clean <- corrPrune(longitudinal_predictors, threshold = 0.7)
+
+# Model-based pruning
+longitudinal_model <- modelPrune(
+  formula = outcome ~ .,
+  data = cbind(
+    outcome = longitudinal_example$outcome,
+    longitudinal_clean
+  ),
+  limit = 5
+)
+
+# Save results
+saveRDS(list(
+  longitudinal_clean = longitudinal_clean,
+  longitudinal_model = longitudinal_model,
+  original_dim = dim(longitudinal_predictors)
[…]
