### 2025-11-25 f818ed8
Enable caching for workflows vignette to speed up builds
Added cache=TRUE to workflows.Rmd setup chunk. This will cache
all computations, dramatically speeding up:
- pkgdown::build_site()
- R CMD check
- devtools::build_vignettes()

First build will still be slow, but subsequent builds will
use cached results and complete in seconds instead of minutes.

Cache directory (vignettes/workflows_cache/) is git-ignored.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
