+++ vignettes/workflows.Rmd
@@ -14,7 +14,8 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.width = 7,
-  fig.height = 5
+  fig.height = 5,
+  cache = TRUE  # Cache expensive computations to speed up builds
 )
 ```
 
