### 2025-11-25 a1571d8
Fix log-scale plotting in advanced vignette
- Replace NA/zero benchmark values correctly for log scale
- Plot only valid data points (exact mode stops at p=25)
- Update annotation to show 'Exact mode limit' at p=25
- Apply standardized RGB colors (red=exact, blue=greedy)
- Prevents log(0) and log(NA) errors that break axis rendering
