### 2025-11-25 f1f3442
Extend benchmark range to p=300 for comprehensive scaling analysis
- Remove conservative p<=25 cutoff for exact mode benchmarks
- Benchmark both exact and greedy modes up to p=300
- Update p_values: c(10, 20, 50, 100, 200, 300)
- Remove artificial 'Exact mode limit' annotation from plot
- Update insight text to reflect actual scaling behavior
- Demonstrates both modes remain practical for p<=300
