### 2025-11-25 0935a87
Set exact mode cutoff at p=100 for benchmark scalability
- Exact mode: benchmark only for p <= 100
- Greedy mode: benchmark for all p values (10-300)
- Add vertical line at p=100 marking 'Exact mode limit'
- Update insight: exact mode impractical beyond p≈100
- Red line stops at p=100, blue line continues to p=300
- Demonstrates greedy mode's suitability for high-dimensional data
