### 2025-11-25 7a44d88
Benchmark both exact and greedy modes up to p=300
- Remove artificial p<=100 cutoff for exact mode
- Both modes benchmarked for all p values (10, 20, 50, 100, 200, 300)
- Remove 'Exact mode limit' annotation - no longer needed
- Update insight: both modes scale efficiently for typical correlation structures
- Demonstrates package efficiency for moderately correlated data
- Users can observe actual scaling behavior and choose appropriate mode
