### 2025-11-25 7a4da4c
Prepare v3.0.0 for CRAN submission
Major enhancements:
- Add comprehensive alt-text to all vignette figures for accessibility
- Fix Unicode characters in documentation for LaTeX compatibility
- Enhance vignettes with detailed explanations and workflows
- Update cran-comments.md with comprehensive submission notes
- Rebuild pkgdown website with latest updates

Documentation improvements:
- workflows.Rmd: Add detailed explanations for all 4 workflows
- advanced.Rmd: Enhance custom engines section with conceptual guidance
- quickstart.Rmd: Add alt-text to all figures
- comparison.Rmd: Add alt-tex
[…]
