+++ docs/articles/advanced.md
@@ -1,0 +1,1337 @@
+# Advanced Topics
+
+### Overview
+
+This vignette covers advanced topics for power users, researchers, and
+method developers:
+
+1.  **Understanding the Algorithms** - Exact vs greedy, complexity
+    analysis
+2.  **Custom Engines** - Integrate any modeling package (INLA, mgcv,
+    brms)
+3.  **Exact Subset Enumeration** - Multiple maximal subsets
+4.  **Performance Optimization** - Speed and memory considerations
+5.  **Troubleshooting** - Common issues and solutions
+
+**Target audience**: Users comfortable with R programming and
+statistical methods
+
+**Estimated time**: 15-20 minutes
+
+------------------------------------------------------------------------
+
+## 1. Understanding the Algorithms
+
+### 1.1 Exact vs Greedy: When to Use Each
+
+corrselect offers two algorithmic approaches for
+[`corrPrune()`](https://gcol33.github.io/corrselect/reference/corrPrune.md):
+
+#### Exact Mode (Graph-Theoretic)
+
+**Algorithm**: Eppstein–Löffler–Strash (ELS) or Bron–Kerbosch
+**Complexity**: O(2^p) - exponential in number of predictors
+**Guarantee**: Finds the **largest** maximal independent set
+
+``` r
+
+library(corrselect)
+data(mtcars)
+
+# Exact mode: guaranteed optimal
+exact_result <- corrPrune(mtcars, threshold = 0.7, mode = "exact")
+cat("Exact mode kept:", ncol(exact_result), "variables\n")
+#> Exact mode kept: 5 variables
+```
+
+**Use exact mode when**:
+
+- p ≤ 20 (feasible runtime)
+- You need guaranteed optimal solution
+- Reproducibility is critical
+- You’re writing a paper (justify optimality)
+
+#### Greedy Mode (Heuristic)
+
+**Algorithm**: Deterministic iterative removal **Complexity**: O(p² × k)
+where k = iterations **Guarantee**: Near-optimal, deterministic
+
+``` r
+
+# Greedy mode: fast approximation
+greedy_result <- corrPrune(mtcars, threshold = 0.7, mode = "greedy")
+cat("Greedy mode kept:", ncol(greedy_result), "variables\n")
+#> Greedy mode kept: 5 variables
+```
+
+**Use greedy mode when**:
+
+- p \> 20 (exact becomes slow)
+- Speed is priority
+- Near-optimal is acceptable
+- High-dimensional data (p \> 100)
+
+#### Auto Mode (Recommended)
+
+Automatically selects based on p:
+
+``` r
+
+# Auto mode: smart switching (exact if p ≤ 20, greedy otherwise)
+auto_result <- corrPrune(mtcars, threshold = 0.7, mode = "auto")
+cat("Auto mode kept:", ncol(auto_result), "variables\n")
+#> Auto mode kept: 5 variables
+```
+
+### 1.2 Complexity Analysis with Benchmarks
+
+Let’s measure runtime scaling:
+
+``` r
+
+# Generate datasets with increasing p
+library(microbenchmark)
+
+benchmark_corrPrune <- function(p_values) {
+  results <- data.frame(
+    p = integer(),
+    exact_time_ms = numeric(),
+    greedy_time_ms = numeric()
+  )
+
+  for (p in p_values) {
+    # Generate correlated data
+    set.seed(123)
+    cor_mat <- 0.5^abs(outer(1:p, 1:p, "-"))
+    data <- as.data.frame(MASS::mvrnorm(n = 100, mu = rep(0, p), Sigma = cor_mat))
+
+    # Exact mode (skip if p too large)
+    exact_time_ms <- if (p <= 500) {
[…]
