### 2025-11-25 3020266
Fix Unicode characters and DOI references for CRAN submission
CRAN readiness improvements:
- Replace Unicode characters (τ, ≈, ≤) with LaTeX-safe equivalents in R documentation
- Correct Mahadev & Peled (1995) DOI reference to accurate citation
- Add DOI verification scripts for quality assurance
- Update CLAUDE.md with DOI handling guidelines

Documentation fixes:
- R/corrPrune.R: Replace ≤ with <= in parameter documentation
- R/cor_example.R: Replace τ with "tau" throughout
- R/data.R: Replace ≈ with ~= in dataset descriptions
- R/modelPrune.R: Replace ≤ with <= in algorithm description
- vi
[…]
