+++ R/modelPrune.R
@@ -55,7 +55,7 @@
 #' 3. Computing diagnostics for each fixed-effect predictor
 #' 4. Checking feasibility of `force_in` constraints
 #' 5. Iteratively removing the predictor with the worst diagnostic value
-#'    (excluding `force_in` variables) until all diagnostics ≤ `limit`
+#'    (excluding `force_in` variables) until all diagnostics <= `limit`
 #' 6. Returning the pruned data frame
 #'
 #' **Random Effects**: For mixed models (lme4, glmmTMB), only fixed-effect
