+++ man/genes_example.Rd
@@ -26,10 +26,10 @@ organized into co-expression modules with a binary disease outcome.
 This dataset simulates a high-dimensional, low-sample scenario common in genomics.
 Genes are organized into four co-expression modules:
 \itemize{
-\item Module 1 (GENE001-GENE050): Highly correlated (r ≈ 0.80), disease-associated
-\item Module 2 (GENE051-GENE100): Moderately correlated (r ≈ 0.60)
-\item Module 3 (GENE101-GENE150): Weakly correlated (r ≈ 0.40)
-\item Module 4 (GENE151-GENE200): Independent (r ≈ 0)
+\item Module 1 (GENE001-GENE050): Highly correlated (r ~= 0.80), disease-associated
+\item Module 2 (GENE051-GENE100): Moderately correlated (r ~= 0.60)
+\item Module 3 (GENE101-GENE150): Weakly correlated (r ~= 0.40)
+\item Module 4 (GENE151-GENE200): Independent (r ~= 0)
 }
 
 Disease outcome depends on a subset of genes from Module 1.
