### 2025-11-25 c37297d
Update pkgdown theme to sandstone with dark navbar
