+++ corrselect.Rcheck/00_pkg_src/corrselect/inst/doc/theory.R
@@ -1,0 +1,248 @@
+## ----setup, include = FALSE---------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 7,
+  fig.height = 5
+)
+
+## -----------------------------------------------------------------------------
+library(corrselect)
+
+# Construct a simple 4x4 correlation matrix
+cor_4var <- matrix(c(
+  1.00, 0.85, 0.10, 0.15,
+  0.85, 1.00, 0.12, 0.18,
+  0.10, 0.12, 1.00, 0.75,
+  0.15, 0.18, 0.75, 1.00
+), nrow = 4, byrow = TRUE)
+
+colnames(cor_4var) <- rownames(cor_4var) <- paste0("V", 1:4)
+
+# Display matrix
+print(cor_4var)
+
+## ----fig.width=6, fig.height=4, fig.alt="Text output showing adjacency matrix for threshold graph with 4 variables. Matrix displays 1 where edges exist (variables with absolute correlation below 0.7 can coexist) and 0 where edges don't exist (variables with absolute correlation above 0.7 cannot coexist). The pattern reveals which variable pairs are compatible for inclusion in the same maximal subset."----
+# Adjacency matrix for threshold graph (edges where |cor| < 0.7)
+adj_matrix <- abs(cor_4var) < 0.7
+diag(adj_matrix) <- FALSE  # No self-loops
+
+# Visualize as adjacency matrix
+cat("Threshold graph edges (1 = edge exists):\n")
+print(adj_matrix * 1)
+
+## ----fig.width=8, fig.height=6, fig.alt="Graph visualization with 4 nodes (V1, V2, V3, V4) arranged in a square pattern. Blue edges connect variable pairs with absolute correlation below 0.7 threshold. Red labels on edges show actual correlation values. Large red circles highlight nodes, with white labels. The graph structure reveals two maximal cliques: {V1,V3} and {V2,V4}, corresponding to maximal subsets where all pairwise correlations remain below threshold."----
+# Node positions (arranged in a square for clarity)
+node_pos <- matrix(c(
+  0, 1,    # V1 (top-left)
+  2, 1,    # V2 (top-right)
+  0, 0,    # V3 (bottom-left)
+  2, 0     # V4 (bottom-right)
+), ncol = 2, byrow = TRUE)
+
+# Plot setup
+par(mar = c(2, 2, 3, 2))
+plot(node_pos, type = "n", xlim = c(-0.5, 2.5), ylim = c(-0.5, 1.5),
+     xlab = "", ylab = "", axes = FALSE,
+     main = "Threshold Graph (τ = 0.7)\nEdges connect variables with |correlation| < 0.7")
+
+# Draw edges (where correlation < 0.7)
+edge_color <- rgb(0.2, 0.5, 0.8, 0.6)
+edge_lwd <- 2
+
+for (i in 1:4) {
+  for (j in 1:4) {
+    if (i < j && adj_matrix[i, j]) {
+      segments(node_pos[i, 1], node_pos[i, 2],
+               node_pos[j, 1], node_pos[j, 2],
+               col = edge_color, lwd = edge_lwd)
+    }
+  }
+}
+
+# Draw nodes
+node_size <- 0.15
+for (i in 1:4) {
+  # Node circle
+  symbols(node_pos[i, 1], node_pos[i, 2],
+          circles = node_size, add = TRUE,
+          inches = FALSE, bg = "white", fg = "black", lwd = 2)
+
+  # Node label
+  text(node_pos[i, 1], node_pos[i, 2],
+       labels = paste0("V", i), cex = 1.2, font = 2)
+}
+
[…]
