+++ corrselect.Rcheck/00_pkg_src/corrselect/R/RcppExports.R
@@ -1,0 +1,23 @@
+# Generated by using Rcpp::compileAttributes() -> do not edit by hand
+# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+findAllMaxSets <- function(corMatrix, threshold, method = "els", force_in = NULL, use_pivot = NULL) {
+    .Call(`_corrselect_findAllMaxSets`, corMatrix, threshold, method, force_in, use_pivot)
+}
+
+runBronKerbosch <- function(corMatrix, threshold, forcedVec, usePivot) {
+    .Call(`_corrselect_runBronKerbosch`, corMatrix, threshold, forcedVec, usePivot)
+}
+
+runELS <- function(corMatrix, threshold, forcedVec) {
+    .Call(`_corrselect_runELS`, corMatrix, threshold, forcedVec)
+}
+
+greedyPruneBackend <- function(assoc_matrix, threshold, force_in = NULL) {
+    .Call(`_corrselect_greedyPruneBackend`, assoc_matrix, threshold, force_in)
+}
+
+runSpectral <- function(corMatrix, threshold, forcedVec, lengthLimit, k_param = NULL) {
+    .Call(`_corrselect_runSpectral`, corMatrix, threshold, forcedVec, lengthLimit, k_param)
+}
+

[…]
