### 2025-11-25 425deff
Add dark navbar CSS
