+++ corrselect.Rcheck/corrselect/doc/comparison.R
@@ -1,0 +1,322 @@
+## ----setup, include = FALSE---------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width = 8,
+  fig.height = 5
+)
+library(corrselect)
+
+## -----------------------------------------------------------------------------
+data(bioclim_example)
+predictors <- bioclim_example[, -1]  # Exclude response
+response <- bioclim_example[, 1]
+
+cat("Variables:", ncol(predictors), "\n")
+cat("Observations:", nrow(predictors), "\n")
+cat("Response: species_richness (continuous)\n")
+
+## ----fig.width=8, fig.height=6, fig.alt="Correlation heatmap of 19 bioclimatic variables displayed as a color-coded matrix. Blue indicates negative correlations, white indicates near-zero correlations, and red indicates positive correlations. Numerical correlation values are overlaid on each cell. The heatmap reveals block structure with correlations ranging from -0.15 to 0.97, showing strong correlations among temperature-related variables and precipitation-related variables."----
+cor_matrix <- cor(predictors)
+
+# Correlation heatmap
+col_pal <- colorRampPalette(c("#3B4992", "white", "#EE0000"))(100)
+
+par(mar = c(1, 1, 3, 1))
+image(1:ncol(cor_matrix), 1:nrow(cor_matrix), t(cor_matrix[nrow(cor_matrix):1, ]),
+      col = col_pal,
+      xlab = "", ylab = "", axes = FALSE,
+      main = "Bioclimatic Variable Correlations (p = 19)",
+      zlim = c(-1, 1))
+axis(1, at = 1:ncol(cor_matrix), labels = colnames(cor_matrix), las = 2, cex.axis = 0.7)
+axis(2, at = ncol(cor_matrix):1, labels = colnames(cor_matrix), las = 2, cex.axis = 0.7)
+
+for (i in 1:ncol(cor_matrix)) {
+  for (j in 1:nrow(cor_matrix)) {
+    text_col <- if (abs(cor_matrix[j, i]) > 0.6) "white" else "black"
+    text(i, nrow(cor_matrix) - j + 1, sprintf("%.2f", cor_matrix[j, i]),
+         cex = 0.5, col = text_col)
+  }
+}
+
+## -----------------------------------------------------------------------------
+if (requireNamespace("caret", quietly = TRUE)) {
+  # Apply caret's greedy algorithm
+  to_remove_caret <- caret::findCorrelation(cor_matrix, cutoff = 0.7)
+  result_caret <- predictors[, -to_remove_caret]
+
+  cat("caret results:\n")
+  cat("  Variables retained:", ncol(result_caret), "\n")
+  cat("  Variables removed:", length(to_remove_caret), "\n")
+  cat("  Removed:", paste(colnames(predictors)[to_remove_caret], collapse = ", "), "\n")
+}
+
+## -----------------------------------------------------------------------------
+# Apply corrselect (exact mode)
+result_corrselect <- corrPrune(predictors, threshold = 0.7, mode = "exact")
+
+cat("\ncorrselect results:\n")
+cat("  Variables retained:", ncol(result_corrselect), "\n")
+cat("  Variables removed:", length(attr(result_corrselect, "removed_vars")), "\n")
+cat("  Removed:", paste(attr(result_corrselect, "removed_vars"), collapse = ", "), "\n")
+
[…]
