+++ corrselect.Rcheck/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(corrselect)
+
+test_check("corrselect")
