### 2025-11-25 31a24a3
Update pkgdown theme to sandstone with light-switch, simplify quickstart vignette
- Use standard sandstone bootswatch theme
- Add light-switch toggle for dark/light mode
- Simplify quickstart vignette to use mtcars instead of bioclim_example
- Add extra.css for plot background colors in both themes

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
