+++ docs/articles/quickstart.md
@@ -5,7 +5,7 @@
 corrselect identifies and removes redundant variables based on pairwise
 correlation or association. Given a threshold $`\tau`$, it finds subsets
 where all pairwise associations satisfy $`|a_{ij}| < \tau`$ (see
-[`vignette("theory")`](https://gcol33.github.io/corrselect/articles/theory.md)
+[`vignette("theory")`](https://gillescolling.com/corrselect/articles/theory.md)
 for mathematical formulation).
 
 ## Interface Hierarchy
@@ -32,7 +32,7 @@ correlation:
 **corrSelect()** - Returns all maximal subsets (numeric data):
 
 - Enumerates all maximal valid subsets satisfying threshold (see
-  [`vignette("theory")`](https://gcol33.github.io/corrselect/articles/theory.md))
+  [`vignette("theory")`](https://gillescolling.com/corrselect/articles/theory.md))
 - Provides full metadata (size, avg_corr, max_corr, min_corr)
 - Exact or greedy search
 
@@ -52,24 +52,23 @@ correlation:
 
 ------------------------------------------------------------------------
 
-## Pruning Example: corrPrune()
+## Quick Examples
+
+### corrPrune(): Association-Based Pruning
 
 ``` r
 
 library(corrselect)
-data(bioclim_example)
+data(mtcars)
 
 # Remove correlated predictors (threshold = 0.7)
-pruned <- corrPrune(
-  data = bioclim_example[, -1],  # Exclude response
-  threshold = 0.7
-)
+pruned <- corrPrune(mtcars, threshold = 0.7)
 
-# Variables retained
-ncol(pruned)
-#> [1] 12
-head(names(pruned))
-#> [1] "BIO1"  "BIO3"  "BIO6"  "BIO9"  "BIO11" "BIO12"
+# Results
+cat(sprintf("Reduced from %d to %d variables\n", ncol(mtcars), ncol(pruned)))
+#> Reduced from 11 to 5 variables
+names(pruned)
+#> [1] "mpg"  "drat" "qsec" "gear" "carb"
 ```
 
 Variables removed:
@@ -77,13 +76,13 @@ Variables removed:
 ``` r
 
 attr(pruned, "removed_vars")
-#> [1] "BIO2"  "BIO4"  "BIO5"  "BIO7"  "BIO8"  "BIO10" "BIO15"
+#> [1] "cyl"  "disp" "hp"   "wt"   "vs"   "am"
 ```
 
 **How corrPrune() selects among multiple maximal subsets**:
 
 When multiple maximal subsets exist (which is common),
-[`corrPrune()`](https://gcol33.github.io/corrselect/reference/corrPrune.md)
+[`corrPrune()`](https://gillescolling.com/corrselect/reference/corrPrune.md)
 returns the subset with the **lowest average absolute correlation**.
 This selection criterion balances three goals:
 
@@ -94,193 +93,52 @@ This selection criterion balances three goals:
 3.  **Deterministic behavior**: Always returns the same result for the
     same data
 
-**Example**:
-
-``` r
-
-# If two maximal subsets exist:
-# Subset A: {bio1, bio4, bio12} with avg_corr = 0.35
-# Subset B: {bio2, bio5, bio15} with avg_corr = 0.42
-# corrPrune() returns Subset A (lower average correlation)
-```
-
 To explore **all** maximal subsets instead of just the optimal one, use
-[`corrSelect()`](https://gcol33.github.io/corrselect/reference/corrSelect.md)
+[`corrSelect()`](https://gillescolling.com/corrselect/reference/corrSelect.md)
 (see below).
 
-Visualization:
-
-``` r
-
-cor_before <- cor(bioclim_example[, -1])
[…]
