### 2025-11-25 e9fbf7d
v3.0.1: SVG output, dark mode support, updated defaults
- Switch vignettes to svglite for SVG figure output
- Add CSS to recolor white SVG fills in dark mode
- Add boxed legends with white background across all vignettes
- Increase max_exact_p default from 20 to 100 in corrPrune()
- Fix manual VIF removal example in comparison vignette
- Simplify threshold selection section in quickstart vignette

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
