+++ vignettes/theory.Rmd
@@ -14,7 +14,8 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.width = 7,
-  fig.height = 5
+  fig.height = 5,
+  dev = "svglite"
 )
 ```
 
@@ -332,7 +333,7 @@ text(1, -0.35, "cor = 0.75\n(too high, no edge)", cex = 0.8, col = "red")
 # Legend showing maximal cliques
 legend("right",
        legend = c("Maximal cliques:", "{V1, V3}", "{V1, V4}", "{V2, V3}", "{V2, V4}"),
-       bty = "n", cex = 0.9,
+       bty = "o", bg = "white", cex = 0.9,
        pch = c(NA, 19, 19, 19, 19),
        col = c(NA, "black", "black", "black", "black"))
 
@@ -405,7 +406,7 @@ if (requireNamespace("igraph", quietly = TRUE)) {
                    "Block 3 (V11-V15): Low cor",
                    "Block 4 (V16-V20): Minimal cor"),
          fill = c("#d73027", "#fc8d59", "#91bfdb", "#4575b4"),
-         bty = "n", cex = 0.8)
+         bty = "o", bg = "white", cex = 0.8)
 } else {
   cat("Install igraph for network visualization: install.packages('igraph')\n")
   cat("Adjacency matrix (first 5×5 block):\n")
@@ -694,7 +695,8 @@ legend("bottomleft",
        lty = c(NA, NA, 1),
        col = c("black", "black", "gray70"),
        lwd = c(2, 2, 1.5),
-       bty = "n",
+       bty = "o",
+       bg = "white",
        cex = 0.7)
 
 par(mfrow = c(1, 1))
