+++ vignettes/advanced.Rmd
@@ -14,7 +14,8 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.width = 7,
-  fig.height = 5
+  fig.height = 5,
+  dev = "svglite"
 )
 ```
 
@@ -185,7 +186,8 @@ legend("topleft",
        col = c(rgb(0.8, 0.2, 0.2, 1), rgb(0.2, 0.5, 0.8, 1)),
        pch = 19,
        lwd = 2,
-       bty = "n")
+       bty = "o",
+       bg = "white")
 ```
 
 **Key insight**: The scaling behavior depends heavily on correlation structure. For this moderately correlated dataset (ρ = 0.5^|i-j|), exact mode remains practical up to p ≈ 500, while greedy mode scales efficiently beyond p = 1000. The exact mode provides guaranteed optimality at the cost of computational complexity, while greedy mode offers substantial speed advantages for large p with near-optimal results in most practical scenarios.
@@ -844,7 +846,8 @@ hist(abs(cor_vec), breaks = 30,
 abline(v = c(0.5, 0.7, 0.9), col = c("red", "blue", "green"), lwd = 2, lty = 2)
 legend("topright",
        legend = c("0.5 (strict)", "0.7 (moderate)", "0.9 (lenient)"),
-       col = c("red", "blue", "green"), lwd = 2, lty = 2)
+       col = c("red", "blue", "green"), lwd = 2, lty = 2,
+       bty = "o", bg = "white")
 
 # Subset size vs threshold
 thresholds <- seq(0.3, 0.95, by = 0.05)

[…]
