### 2025-11-25 5cd5d45
Fix kappa axis ticks using pretty() for automatic scaling
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
