### 2025-11-25 ff1b058
Replace 1:nrow/ncol with seq_len() in vignette, add CSS for image sizing
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
