### 2025-11-25 362938d
Remove svglite device from vignettes for pkgdown compatibility
The svglite device caused "Expected string vector of length 1" errors
when pkgdown passed file paths to it. Using the default PNG device
resolves this and allows pkgdown::build_site() to complete successfully.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
