### 2025-11-25 92ea0f4
Add lightswitch to navbar for dark/light mode toggle
The navbar structure was explicitly defined without the lightswitch
component, so it wasn't appearing despite light-switch: true being set.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
