+++ _pkgdown.yml
@@ -18,7 +18,7 @@ authors:
 navbar:
   structure:
     left:  [intro, reference, articles, news]
-    right: [search, github]
+    right: [search, lightswitch, github]
   components:
     intro:
       text: Get Started
