### 2025-11-25 115674b
Fix dark mode plot visibility with CSS invert filter
Use filter: invert() to make PNG plots readable in dark mode by
inverting colors, turning white backgrounds dark.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
