### 2025-11-25 b167397
Switch to SVG figures with theme-aware backgrounds
- Configure pkgdown to use SVG device for figures
- Add dev = "svg" to vignettes for consistent output
- CSS sets background-color on SVG images matching sandstone theme
- Light mode: #f8f5f0, Dark mode: #3e3f3a

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
