+++ vignettes/workflows.Rmd
@@ -15,6 +15,7 @@ knitr::opts_chunk$set(
   comment = "#>",
   fig.width = 7,
   fig.height = 5,
+  dev = "svg",
   cache = TRUE  # Cache expensive computations to speed up builds
 )
 ```
