### 2025-11-25 2e0a2f9
Rebuild pkgdown site
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
