### 2025-11-25 345f0d7
Fix SVG plot styling for light/dark mode
- Add bg: "transparent" to pkgdown figures config
- Replace bg = "white" with bg = NA in all legend calls
- Add rounded corners to match code blocks
- Add subtle border for visual consistency

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
