+++ vignettes/workflows.Rmd
@@ -133,7 +133,7 @@ legend("topright",
        lwd    = c(NA, NA, 2),
        col    = c(NA, NA, "black"),
        bty    = "o",
-       bg     = "white")
+       bg = NA)
 ```
 
 ### Fit models
@@ -273,7 +273,7 @@ legend(
   lwd    = 2,
   horiz  = TRUE,
   bty    = "o",
-  bg     = "white",
+  bg = NA,
   x.intersp = 0.8
 )
 
@@ -348,7 +348,7 @@ legend(
   fill = c(col_full, col_pruned),
   border = "white",
   bty = "o",
-  bg = "white"
+  bg = NA
 )
 
 ```
@@ -809,7 +809,7 @@ legend(
   lwd = c(NA, NA, 2),
   col = c(NA, NA, "red"),
   bty = "o",
-  bg = "white"
+  bg = NA
 )
 ```
 
@@ -862,3 +862,4 @@ These four workflows demonstrate how **corrselect** integrates into diverse anal
 
 **Methods**:
 - See package documentation and JOSS paper for algorithm details
+
