### 2025-11-25 7bfa5b3
Use base R svg device with transparent bg and CSS styling
- Switch from svglite to base R svg (svglite fails with pkgdown)
- Set bg: NA for transparent SVG backgrounds
- CSS provides background-color matching theme
- Add rounded corners and border to match code blocks
- width: 100% ensures images fill container width

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
