+++ vignettes/advanced.Rmd
@@ -1033,3 +1033,5 @@ pipeline <- function(data, response) {
 - `?modelPrune` - Model-based pruning
 - `?corrSelect` - Exact subset enumeration
 
+
+
