+++ vignettes/comparison.Rmd
@@ -14,8 +14,7 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.width = 8,
-  fig.height = 5,
-  dev = "svg"
+  fig.height = 5
 )
 library(corrselect)
 ```
@@ -590,4 +589,7 @@ final_model <- lm(response ~ ., data = model_data[, c("response", important_vars
 - `vignette("workflows")` - Domain-specific workflows (genomics, ecology, surveys)
 - `vignette("advanced")` - Algorithm selection and performance tuning
 - `vignette("theory")` - Graph-theoretic foundations and formal proofs
+
+
+
 
