### 2025-11-25 c4fa01e
Fix dark mode: invert SVG colors with hue-rotate to preserve data colors
The invert(1) hue-rotate(180deg) filter inverts black to white (and vice versa)
while rotating hue back so data colors remain visually similar.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
