### 2025-11-25 8696e4e
Use Bootstrap CSS variables for plot backgrounds
Use --bs-tertiary-bg and --bs-border-color to match code block styling
automatically in both light and dark modes.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
