### 2025-11-25 679ffe2
Remove filter, use simple CSS background for plots
Use --bs-tertiary-bg for background in both modes. Plot content (text, lines)
remains black but background matches theme.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
