+++ docs/articles/advanced.md
@@ -138,13 +138,13 @@ benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
 #> 1   10           0.5            0.2
-#> 2   20           0.6            0.3
-#> 3   50           1.4            0.7
-#> 4  100           4.4            1.4
+#> 2   20           0.7            0.3
+#> 3   50           1.4            0.8
+#> 4  100           4.5            1.4
 #> 5  200          22.6            3.3
-#> 6  300          85.7            6.9
-#> 7  500         310.2           15.5
-#> 8 1000            NA           54.1
+#> 6  300          84.7            6.8
+#> 7  500         308.1           15.1
+#> 8 1000            NA           55.3
 ```
 
 ``` r
@@ -740,9 +740,9 @@ time2 <- median(microbenchmark(
 )$time) / 1e6  # Convert nanoseconds to milliseconds
 
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
-#> Recomputing each time: 4.2 ms
+#> Recomputing each time: 4.0 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
-#> Precomputed matrix: 2.1 ms
+#> Precomputed matrix: 2.0 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
 #> Speedup: 2.0x faster
 ```
