### 2025-11-25 323a165
Add SVG post-processing for transparent backgrounds
- Add build_site.R that builds pkgdown then replaces white fills with
  transparent in all SVG files
- Revert legend bg = NA to bg = "white" in vignettes (post-processor handles it)
- White fills (rgb(100%,100%,100%)) become fill="none" for CSS theming

Usage: Rscript build_site.R

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
