+++ vignettes/workflows.Rmd
@@ -132,7 +132,7 @@ legend("topright",
        lwd    = c(NA, NA, 2),
        col    = c(NA, NA, "black"),
        bty    = "o",
-       bg = NA)
+       bg = "white")
 ```
 
 ### Fit models
@@ -272,7 +272,7 @@ legend(
   lwd    = 2,
   horiz  = TRUE,
   bty    = "o",
-  bg = NA,
+  bg = "white",
   x.intersp = 0.8
 )
 
@@ -347,7 +347,7 @@ legend(
   fill = c(col_full, col_pruned),
   border = "white",
   bty = "o",
-  bg = NA
+  bg = "white"
 )
 
 ```
@@ -808,7 +808,7 @@ legend(
   lwd = c(NA, NA, 2),
   col = c(NA, NA, "red"),
   bty = "o",
-  bg = NA
+  bg = "white"
 )
 ```
 
@@ -861,6 +861,7 @@ These four workflows demonstrate how **corrselect** integrates into diverse anal
 
 **Methods**:
 - See package documentation and JOSS paper for algorithm details
+
 
 
 
