+++ vignettes/comparison.Rmd
@@ -164,7 +164,7 @@ if (requireNamespace("caret", quietly = TRUE)) {
          lwd    = c(NA, NA, NA, 2),
          col    = c(NA, NA, NA, "black"),
          bty    = "o",
-         bg = NA)
+         bg = "white")
 }
 ```
 
@@ -504,7 +504,7 @@ if (requireNamespace("car", quietly = TRUE)) {
          lwd    = c(NA, NA, 2),
          col    = c(NA, NA, "black"),
          bty    = "o",
-         bg = NA)
+         bg = "white")
 }
 ```
 
@@ -589,6 +589,7 @@ final_model <- lm(response ~ ., data = model_data[, c("response", important_vars
 - `vignette("workflows")` - Domain-specific workflows (genomics, ecology, surveys)
 - `vignette("advanced")` - Algorithm selection and performance tuning
 - `vignette("theory")` - Graph-theoretic foundations and formal proofs
+
 
 
 
