### 2025-11-25 632846e
Exclude build_site.R from package and git
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
