### 2025-11-25 fc10a57
Theme-aware SVG plots for light/dark mode support
- Update build_site.R to inject CSS media queries into SVGs
- SVG backgrounds now auto-switch between light (#ffffff) and dark (#212529)
- Simplify extra.css (no background/border needed, SVGs self-style)
- Plot and legend backgrounds adapt to prefers-color-scheme

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
