### 2025-11-25 b441762
Add theme-aware text styling to SVG plots
- Update build_site.R to replace black fills with .theme-text class
- Inject font-family (Roboto) and text colors matching Sandstone theme
- Light mode text: #3e3f3a, Dark mode text: #dfd7ca
- All 16 SVG plots now have theme-aware backgrounds AND text

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
