+++ docs/articles/advanced.md
@@ -137,14 +137,14 @@ p_values <- c(10, 20, 50, 100, 200, 300, 500, 1000)
 benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
-#> 1   10           0.5            0.3
-#> 2   20           0.6            0.6
-#> 3   50           1.4            1.0
-#> 4  100           4.4            1.4
-#> 5  200          22.6            3.3
-#> 6  300          86.2            7.0
-#> 7  500         312.7           15.8
-#> 8 1000            NA           54.7
+#> 1   10           0.6            0.2
+#> 2   20           0.6            0.4
+#> 3   50           1.4            0.8
+#> 4  100           4.6            1.6
+#> 5  200          23.0            3.4
+#> 6  300          90.3            7.6
+#> 7  500         314.8           15.7
+#> 8 1000            NA           56.6
 ```
 
 ``` r
@@ -740,11 +740,11 @@ time2 <- median(microbenchmark(
 )$time) / 1e6  # Convert nanoseconds to milliseconds
 
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
-#> Recomputing each time: 4.3 ms
+#> Recomputing each time: 4.0 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
-#> Precomputed matrix: 2.2 ms
+#> Precomputed matrix: 2.0 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
-#> Speedup: 1.9x faster
+#> Speedup: 2.0x faster
 ```
 
 **Use precomputed matrices when**:
