### 2025-11-25 2d85c07
Fix Sandstone theme colors for SVG plots
Correct the dark mode colors to match actual Bootswatch Sandstone values:
- Light: bg=#FFFFFF, text=#3E3F3A
- Dark:  bg=#3E3F3A, text=#E9E7E3

Previous values incorrectly used Bootstrap default dark (#212529) instead
of Sandstone's dark background (#3E3F3A).

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
