### 2025-11-25 6c729fb
Add code-block style borders to SVG plots
Apply matching borders from Sandstone theme code blocks:
- Light mode: #DAD3C3
- Dark mode: #4A4B46
- 0.25rem border-radius for consistency

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
