+++ docs/articles/advanced.md
@@ -137,14 +137,14 @@ p_values <- c(10, 20, 50, 100, 200, 300, 500, 1000)
 benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
-#> 1   10           0.5            0.2
+#> 1   10           0.5            0.3
 #> 2   20           0.6            0.4
-#> 3   50           1.5            0.7
-#> 4  100           4.4            1.4
-#> 5  200          22.5            3.3
-#> 6  300          86.4            6.5
-#> 7  500         311.0           15.6
-#> 8 1000            NA           56.5
+#> 3   50           1.4            0.8
+#> 4  100           4.5            1.4
+#> 5  200          22.6            3.4
+#> 6  300          86.3            7.0
+#> 7  500         312.6           15.5
+#> 8 1000            NA           59.0
 ```
 
 ``` r
@@ -740,9 +740,9 @@ time2 <- median(microbenchmark(
 )$time) / 1e6  # Convert nanoseconds to milliseconds
 
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
-#> Recomputing each time: 4.2 ms
+#> Recomputing each time: 4.4 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
-#> Precomputed matrix: 1.9 ms
+#> Precomputed matrix: 2.0 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
 #> Speedup: 2.2x faster
 ```
