### 2025-11-25 f7f5f52
Inline SVGs in HTML for proper dark mode toggle support
- Add inline_svgs_in_html() to replace <img src="*.svg"> with inline <svg>
- Inline SVGs can be styled by page CSS (data-bs-theme attribute)
- Add dark mode CSS rules for .theme-bg and .theme-text classes
- SVGs now respond to the page's light/dark toggle, not just system preference

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
