### 2025-11-26 a5a038e
Fix dark mode CSS specificity with !important
The embedded SVG <style> blocks were overriding external CSS.
Add !important to dark mode fill rules to ensure proper override.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
