### 2025-11-26 bef9511
Fix dark mode text color to match Sandstone body-color
Use correct Sandstone dark mode text color #dfd7ca (was #E9E7E3)
- Light mode: #3e3f3a (unchanged)
- Dark mode: #dfd7ca (fixed)

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
