### 2025-11-26 3a65e42
Revert to base R svg() after svglite compatibility issues
svglite::svglite caused "Expected string vector of length 1" errors
with pkgdown's figure rendering. Staying with base R svg() device.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
