### 2025-11-26 57778a3
Add theme-aware histogram bar borders for dark mode
- Updated extra.css with CSS rules for dark mode rect strokes
- Histogram bar borders now match text color in each theme
- Legend box dark mode styling partially working (fill/stroke override)

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
