+++ docs/articles/workflows.md
@@ -696,7 +696,7 @@ system.time({
   )
 })
 #>    user  system elapsed 
-#>    0.01    0.00    0.02
+#>    0.02    0.00    0.02
 
 # Reduction
 cat(sprintf("Reduced from %d → %d genes\n",
@@ -765,11 +765,11 @@ greedy_result <- corrPrune(gene_subset, threshold = 0.8, mode = "greedy")
 
 # Compare
 cat(sprintf("Exact mode: %d genes kept (%.1f ms)\n", ncol(exact_result), exact_time))
-#> Exact mode: 11 genes kept (3.3 ms)
+#> Exact mode: 11 genes kept (3.5 ms)
 cat(sprintf("Greedy mode: %d genes kept (%.1f ms)\n", ncol(greedy_result), greedy_time))
 #> Greedy mode: 10 genes kept (0.5 ms)
 cat(sprintf("Speedup: %.1fx faster\n", exact_time / greedy_time))
-#> Speedup: 6.3x faster
+#> Speedup: 6.8x faster
 ```
 
 The greedy mode is substantially faster. For the full 200-gene dataset,
