### 2025-11-26 b74f0bc
Fix legend background theming in SVG post-processing
- Add detection for CSS variable fills (--svg-bg-color) in early-exit check
- Convert CSS variable fills back to #FFFFFF for pattern matching
- Detect legend rects with white fills that need theming
- Properly add stroke and fill to legend boxes without duplicates

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
