### 2025-11-26 c4f3d65
Improve SVG styling: rounded legend boxes, matching border radius
- Match image border-radius to code blocks (.375rem)
- Remove SVG internal stroke, use CSS border for proper rounding
- Add rx='4' to legend boxes for rounded corners
- Exclude stroke:none rects from dark mode stroke override
- Fixes empty legend key boxes showing unwanted borders in dark mode

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
