### 2025-11-26 583d253
Fix SVG background stroke bleeding in dark mode
- Add explicit stroke:none !important for .svg-bg class in CSS
- Exclude .svg-bg from legend box stroke rule using :not(.svg-bg)
- Add overflow:hidden and display:block to SVG container for cleaner borders
- Fix CSS specificity issues causing inner border artifacts

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
