+++ docs/articles/quickstart.md
@@ -1,5 +1,25 @@
 # Quick Start
 
+## Installation
+
+``` r
+
+# Install from CRAN
+install.packages("corrselect")
+
+# Or install development version from GitHub
+# install.packages("devtools")
+devtools::install_github("GillesColling/corrselect")
+```
+
+**Suggested packages** (for extended functionality):
+
+- `lme4`, `glmmTMB`: Mixed-effects models in
+  [`modelPrune()`](https://gillescolling.com/corrselect/reference/modelPrune.md)
+- `WGCNA`: Biweight midcorrelation (`bicor`)
+- `energy`: Distance correlation
+- `minerva`: Maximal information coefficient
+
 ## What corrselect Does
 
 corrselect identifies and removes redundant variables based on pairwise
@@ -50,15 +70,12 @@ correlation:
 - No data preprocessing
 - Useful for repeated analyses
 
-------------------------------------------------------------------------
-
 ## Quick Examples
 
 ### corrPrune(): Association-Based Pruning
 
 ``` r
 
-library(corrselect)
 data(mtcars)
 
 # Remove correlated predictors (threshold = 0.7)
@@ -126,7 +143,7 @@ show(results)
 #>   Method:      bron-kerbosch
 #>   Correlation: pearson
 #>   Threshold:   0.700
-#>   Subsets:     15 valid combinations
+#>   Subsets:     15 maximal subsets
 #>   Data Rows:   32 used in correlation
 #>   Pivot:       TRUE
 #> 
@@ -185,7 +202,7 @@ show(results_mixed)
 #>   AssocMethod: numeric_numeric = pearson, numeric_factor = eta, numeric_ordered
 #>                = spearman, factor_ordered = cramersv
 #>   Threshold:   0.500
-#>   Subsets:     1 valid combinations
+#>   Subsets:     1 maximal subsets
 #>   Data Rows:   100 used in correlation
 #>   Pivot:       TRUE
 #> 
@@ -193,9 +210,11 @@ show(results_mixed)
 #>   No.  Variables                          Avg    Max    Size
 #>   ------------------------------------------------------------
 #>   [ 1] x1, x2, cat1, ord1                0.077  0.184     4
-```
 
-------------------------------------------------------------------------
+# Verify all pairwise associations are below threshold
+cat("Max pairwise association:", max(results_mixed@max_corr), "\n")
+#> Max pairwise association: 0.1842642
+```
 
 ## Protecting Variables
 
@@ -215,23 +234,18 @@ pruned_force <- corrPrune(
 #> [1] TRUE
 ```
 
-------------------------------------------------------------------------
-
 ## Threshold Selection
 
-Common threshold guidelines:
+Common thresholds: **0.5** (strict), **0.7** (moderate, recommended
+default), **0.9** (lenient).
 
-| Threshold | Interpretation | Use Case                             |
-|-----------|----------------|--------------------------------------|
-| 0.5       | Strict         | When multicollinearity is critical   |
-| 0.7       | Moderate       | General-purpose default              |
-| 0.9       | Lenient        | When retaining variables is priority |
+Lower thresholds are stricter because they allow fewer variable pairs to
+coexist, resulting in smaller subsets. Higher thresholds permit stronger
+correlations, retaining more variables.
 
 For detailed threshold selection strategies including visualization
-techniques and sensitivity analysis, see
+techniques, VIF guidelines, and sensitivity analysis, see
[…]
