### 2025-11-26 aa49110
Bump to v3.0.2: single-quote package names per CRAN policy
- Single-quoted 'lme4' and 'glmmTMB' in DESCRIPTION
- Updated NEWS.md

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
