### 2025-11-26 d655165
Update cran-comments.md and CRAN-SUBMISSION for v3.0.2
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
