### 2025-11-26 2dd7fe2
Improve NEWS.md: add 3.0.1 and 2.0.0 sections, reorganize 3.0.0
- Move bug fixes from 3.0.0 to new 3.0.1 section
- Add 2.0.0 section documenting assocSelect() introduction
- Better historical changelog structure

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
