### 2025-11-26 8691aff
Update NEWS.md 3.0.2 to include vignette changes
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
