### 2025-11-26 6229ee4
Replace em-dashes with colons in theory vignette TOC
