+++ vignettes/theory.Rmd
@@ -14,13 +14,29 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
   fig.width = 7,
-  fig.height = 5
+  fig.height = 5,
+  dev = "svglite",
+  fig.ext = "svg"
 )
+library(corrselect)
 ```
 
 ## Overview
 
 This vignette presents the mathematical formulation and graph-theoretic foundations underlying **corrselect**. Variable subset selection under correlation constraints is formulated as a maximal independent set problem on threshold graphs, enabling exact enumeration via established algorithms from computational graph theory. The vignette defines the formal problem statement, explains the graph-theoretic representation, details the three implemented algorithms (Bron-Kerbosch, Eppstein-Löffler-Strash, and greedy heuristic), analyzes their computational complexity, and provides comprehensive references to the theoretical literature. For practical usage examples, see `vignette("quickstart")` and `vignette("workflows")`. For algorithmic control and performance tuning, see `vignette("advanced")`.
+
+### Contents
+
+1. [Terminology](#terminology): Core definitions (association, threshold, valid subset, clique)
+2. [Intuitive Overview](#intuitive-overview): Conceptual introduction with toy examples
+3. [Problem Formulation](#problem-formulation): Formal mathematical statement
+4. [Graph-Theoretic Interpretation](#graph-theoretic-interpretation): Threshold graphs and maximal cliques
+5. [From Theory to Implementation](#from-theory-to-implementation): How concepts map to function arguments
+6. [Search Algorithms](#search-algorithms): Exact enumeration vs greedy heuristic
+7. [Algorithm Pseudocode](#algorithm-pseudocode): ELS, Bron-Kerbosch, Greedy
+8. [Technical Details](#forced-variables): Forced variables, complexity, output structure
+9. [Design Philosophy](#design-philosophy): Why maximal? Why hard threshold? Why graphs?
+10. [References](#references): Academic literature and further reading
 
 ---
 
@@ -157,6 +173,15 @@ Chooses the method automatically:
 
 This balances optimality with computational cost.
 
+> **Key Points — Terminology**
+>
+> - **Association matrix**: Symmetric matrix of pairwise relationships (correlations, Cramér's V, etc.)
+> - **Threshold (τ)**: Cutoff determining which pairs are "too associated" to coexist
+> - **Valid subset**: All pairwise associations below τ
+> - **Maximal subset**: Valid subset that cannot be enlarged
+> - **Threshold graph**: Graph where edges connect compatible (low-association) variable pairs
+> - **Clique**: Fully connected subgraph; maximal cliques = maximal valid subsets
+
 ---
 
 ## Intuitive Overview
@@ -207,8 +232,6 @@ corrselect finds **all maximal** subsets because:
 Consider 4 variables with this correlation matrix:
 
 ```{r}
-library(corrselect)
-
 # Construct a simple 4x4 correlation matrix
 cor_4var <- matrix(c(
   1.00, 0.85, 0.10, 0.15,
@@ -481,6 +504,14 @@ This toy example shows why corrselect enumerates all solutions:
 
 Real datasets have similar structure but with more variables and more complex clustering.
 
+> **Key Points — Intuitive Overview**
+>
+> - corrselect finds variable subsets where no pair exceeds a correlation threshold
+> - Multiple valid subsets typically exist; all are enumerated for user choice
+> - Graph representation: variables = nodes, low-correlation pairs = edges
+> - Maximal cliques in this graph = maximal valid subsets
+> - Comparing subsets reveals correlation structure (clusters of related variables)
+
 ---
 
[…]
