+++ docs/articles/advanced.md
@@ -138,13 +138,13 @@ benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
 #> 1   10           0.5            0.2
-#> 2   20           0.7            0.4
-#> 3   50           1.4            0.7
-#> 4  100           4.4            1.5
-#> 5  200          22.2            3.3
-#> 6  300          85.1            6.5
-#> 7  500         307.3           15.0
-#> 8 1000            NA           54.0
+#> 2   20           0.8            0.4
+#> 3   50           1.4            0.8
+#> 4  100           4.6            1.4
+#> 5  200          22.8            3.6
+#> 6  300          85.9            6.9
+#> 7  500         308.3           15.5
+#> 8 1000            NA           54.9
 ```
 
 ``` r
@@ -740,11 +740,11 @@ time2 <- median(microbenchmark(
 )$time) / 1e6  # Convert nanoseconds to milliseconds
 
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
-#> Recomputing each time: 4.0 ms
+#> Recomputing each time: 3.9 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
 #> Precomputed matrix: 1.8 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
-#> Speedup: 2.2x faster
+#> Speedup: 2.1x faster
 ```
 
 **Use precomputed matrices when**:
@@ -1363,7 +1363,7 @@ sessionInfo()
 #> [1] stats     graphics  grDevices utils     datasets  methods   base     
 #> 
 #> other attached packages:
-#> [1] microbenchmark_1.5.0 corrselect_3.0.1    
+#> [1] microbenchmark_1.5.0 corrselect_3.0.2    
 #> 
 #> loaded via a namespace (and not attached):
 #>  [1] svglite_2.2.2     cli_3.6.5         knitr_1.50        rlang_1.1.6      
