### 2025-11-26 7a4ab90
Regenerate corrPrune.Rd to fix max_exact_p default value
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
