### 2025-11-27 #8971 @Nikoleta-v3
@jhollist @dansmith01  @danStich👋🏻 Would you be willing to review this submission for JOSS?

As you already know we conduct our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html 😄

The submission I'd like you to review is titled: "corrselect: Exhaustive variable subset selection based on correlation and association matrices". You can find more details at the top of this GitHub issue 🆙
