### 2025-11-28 8dcb9a6
Remove svglite from Suggests (using ragg instead)
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
