+++ DESCRIPTION
@@ -14,7 +14,6 @@ Imports:
     methods,
     stats
 Suggests:
-    svglite,
     GO.db,
     WGCNA,
     preprocessCore,
