### 2025-11-29 2c5beb5
Add responsive table CSS for mobile
Tables now scroll horizontally on mobile instead of wrapping
text awkwardly. Uses white-space: nowrap and overflow-x: auto.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
