+++ docs/news/index.md
@@ -7,16 +7,13 @@
 - Single-quoted software names in DESCRIPTION (‘lme4’, ‘glmmTMB’) per
   CRAN policy
 
+### Documentation
+
+- Updated vignettes with improved examples and workflows
+
 ------------------------------------------------------------------------
 
-## corrselect 3.0.0
-
-### Major Release: Predictor Pruning Toolkit
-
-Version 3.0.0 represents a major expansion of corrselect from a
-specialized subset enumeration tool into a comprehensive predictor
-pruning toolkit. **Fully backward compatible** with 2.x - all existing
-code continues to work.
+## corrselect 3.0.1
 
 ### Bug Fixes
 
@@ -34,6 +31,17 @@ code continues to work.
     for all engines (more robust)
   - Eliminated “Could not find columns” warnings
 - **Test suite**: All 261 tests pass with zero warnings (CRAN-compliant)
+
+------------------------------------------------------------------------
+
+## corrselect 3.0.0
+
+### Major Release: Predictor Pruning Toolkit
+
+Version 3.0.0 represents a major expansion of corrselect from a
+specialized subset enumeration tool into a comprehensive predictor
+pruning toolkit. **Fully backward compatible** with 2.x - all existing
+code continues to work.
 
 ### Major Features
 
@@ -145,3 +153,30 @@ CRAN release: 2025-09-08
 - [`assocSelect()`](https://gillescolling.com/corrselect/reference/assocSelect.md)
   now supports logical columns by automatically converting them to
   factors.
+
+------------------------------------------------------------------------
+
+## corrselect 2.0.0
+
+### Major Release: Mixed-Type Association Selection
+
+Version 2.0.0 introduces support for mixed-type data through the new
+[`assocSelect()`](https://gillescolling.com/corrselect/reference/assocSelect.md)
+function, enabling subset selection on datasets containing numeric,
+factor, and ordered variables.
+
+### Major Features
+
+- **[`assocSelect()`](https://gillescolling.com/corrselect/reference/assocSelect.md)**:
+  New function for mixed-type data frame interface
+  - Handles numeric, factor, and ordered variables
+  - Automatic association measure selection based on variable pair types
+  - Supports Pearson, Spearman, Kendall correlations
+  - Computes Eta-squared for numeric-factor pairs
+  - Computes Cramér’s V for factor-factor pairs
+
+### Enhancements
+
+- Improved algorithm selection logic
+- Better handling of edge cases in subset enumeration
+- Enhanced documentation with examples for mixed-type workflows
