### 2025-11-30 e1af1d0
Rebuild pkgdown site with theme-aware SVGs for dark mode
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
