+++ docs/articles/comparison.md
@@ -742,7 +742,7 @@ sessionInfo()
 #> [1] stats     graphics  grDevices utils     datasets  methods   base     
 #> 
 #> other attached packages:
-#> [1] car_3.1-3        carData_3.0-5    corrselect_3.0.1
+#> [1] car_3.1-3        carData_3.0-5    corrselect_3.0.2
 #> 
 #> loaded via a namespace (and not attached):
 #>  [1] tidyselect_1.2.1     timeDate_4051.111    dplyr_1.1.4         
@@ -758,21 +758,20 @@ sessionInfo()
 #> [31] grid_4.5.1           stats4_4.5.1         future_1.67.0       
 #> [34] ggplot2_4.0.0        globals_0.18.0       scales_1.4.0        
 #> [37] iterators_1.0.14     MASS_7.3-65          cli_3.6.5           
-#> [40] rmarkdown_2.30       ragg_1.5.0           generics_0.1.4      
-#> [43] future.apply_1.20.0  reshape2_1.4.4       cachem_1.1.0        
-#> [46] stringr_1.5.2        splines_4.5.1        parallel_4.5.1      
-#> [49] vctrs_0.6.5          hardhat_1.4.2        glmnet_4.1-10       
-#> [52] Matrix_1.7-4         jsonlite_2.0.0       Formula_1.2-5       
-#> [55] listenv_0.9.1        systemfonts_1.3.1    foreach_1.5.2       
-#> [58] gower_1.0.2          jquerylib_0.1.4      recipes_1.3.1       
-#> [61] glue_1.8.0           parallelly_1.45.1    pkgdown_2.2.0       
-#> [64] codetools_0.2-20     lubridate_1.9.4      stringi_1.8.7       
-#> [67] shape_1.4.6.1        gtable_0.3.6         tibble_3.3.0        
-#> [70] pillar_1.11.1        htmltools_0.5.8.1    ipred_0.9-15        
-#> [73] lava_1.8.2           R6_2.6.1             textshaping_1.0.3   
-#> [76] evaluate_1.0.5       lattice_0.22-7       bslib_0.9.0         
-#> [79] class_7.3-23         Rcpp_1.1.0           svglite_2.2.2       
-#> [82] nlme_3.1-168         prodlim_2025.04.28   ranger_0.17.0       
-#> [85] xfun_0.53            fs_1.6.6             pkgconfig_2.0.3     
-#> [88] ModelMetrics_1.2.2.2
+#> [40] rmarkdown_2.30       generics_0.1.4       future.apply_1.20.0 
+#> [43] reshape2_1.4.4       cachem_1.1.0         stringr_1.5.2       
+#> [46] splines_4.5.1        parallel_4.5.1       vctrs_0.6.5         
+#> [49] hardhat_1.4.2        glmnet_4.1-10        Matrix_1.7-4        
+#> [52] jsonlite_2.0.0       Formula_1.2-5        listenv_0.9.1       
+#> [55] systemfonts_1.3.1    foreach_1.5.2        gower_1.0.2         
+#> [58] jquerylib_0.1.4      recipes_1.3.1        glue_1.8.0          
+#> [61] parallelly_1.45.1    pkgdown_2.2.0        codetools_0.2-20    
+#> [64] lubridate_1.9.4      stringi_1.8.7        shape_1.4.6.1       
+#> [67] gtable_0.3.6         tibble_3.3.0         pillar_1.11.1       
+#> [70] htmltools_0.5.8.1    ipred_0.9-15         lava_1.8.2          
+#> [73] R6_2.6.1             textshaping_1.0.3    evaluate_1.0.5      
[…]
