### 2025-12-04 #9539 @editorialbot
[REVIEW]: corrselect: Exhaustive variable subset selection based on correlation and association matrices
**Submitting author:** @gcol33 (<a href="http://orcid.org/0000-0003-3070-6066">Gilles Colling</a>)
**Repository:** https://github.com/gcol33/corrselect
**Branch with paper.md** (empty if default branch): main
**Version:** v2.0.1
