### 2025-12-04 #9539 @dansmith01
@gcol33, can you confirm that you'd like us to review version 2.0.1 of `corrselect`? I see the development version is 3.0.2.
