### 2025-12-04 gcol33/corrselect#1 @dansmith01
Link to documentation
Can you add a link to the pkgdown documentation website in the "About" section of the github project page's sidebar?

openjournals/joss-reviews#9539
