### 2025-12-04 gcol33/corrselect#3 @dansmith01
Automated cross-platform testing
With `usethis::use_github_action_check_standard()` or similar to ensure that `corrselect` builds without errors across a range of user platforms. Also helps catch bugs before they're released to CRAN.

openjournals/joss-reviews#9539
