### 2025-12-04 gcol33/corrselect#4 @dansmith01
Test coverage
You've got a great set of `testthat` tests. I'd recommend seeing how much of your code they cover with a github action workflow set up by `usethis::use_coverage()`. AI like Claude and Gemini are great at crafting tests to fill in any gaps in coverage.

openjournals/joss-reviews#9539
