### 2025-12-09 #9539 @dansmith01
@gcol33 I'd suggest having us review the most recent version of `corrselect` that satisfies the requirements given in the [review checklist](https://github.com/openjournals/joss-reviews/issues/9539#issuecomment-3612977376), but ultimately it is your decision.
