### 2025-12-09 11eeb0d
Add CRAN downloads badge and improve survey example data
- Add CRAN downloads badge to README
- Modify survey_example dataset to have different within-construct
  correlations (satisfaction high, engagement moderate, loyalty low)
- This makes the vignette visualization more informative by showing
  different retention rates across constructs after pruning
- Rebuild pkgdown site with updated data

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
