### 2025-12-09 e1116b6
Rebuild workflows vignette with fresh cache for new survey data
