+++ docs/articles/advanced.md
@@ -138,13 +138,13 @@ benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
 #> 1   10           0.5            0.3
-#> 2   20           0.8            0.5
-#> 3   50           1.5            0.8
-#> 4  100           4.7            1.6
-#> 5  200          22.9            3.3
-#> 6  300          85.9            6.6
-#> 7  500         307.6           15.2
-#> 8 1000            NA           57.9
+#> 2   20           0.6            0.4
+#> 3   50           1.4            0.7
+#> 4  100           4.4            1.4
+#> 5  200          22.6            3.4
+#> 6  300          85.7            6.8
+#> 7  500         316.1           16.2
+#> 8 1000            NA           58.4
 ```
 
 ``` r
@@ -742,9 +742,9 @@ time2 <- median(microbenchmark(
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
 #> Recomputing each time: 4.1 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
-#> Precomputed matrix: 2.0 ms
+#> Precomputed matrix: 1.9 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
-#> Speedup: 2.0x faster
+#> Speedup: 2.2x faster
 ```
 
 **Use precomputed matrices when**:
