+++ docs/articles/workflows.md
@@ -449,11 +449,11 @@ str(survey_example[, 1:10])  # First 10 columns
 #>  $ gender              : Factor w/ 3 levels "Female","Male",..: 2 3 1 2 2 1 1 2 2 1 ...
 #>  $ education           : Ord.factor w/ 4 levels "High School"<..: 3 1 4 2 2 1 1 1 2 3 ...
 #>  $ overall_satisfaction: num  58 40 44 40 58 67 61 49 51 52 ...
-#>  $ satisfaction_1      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 3 5 3 4 5 5 4 4 5 ...
-#>  $ satisfaction_2      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 3 4 3 4 6 6 4 4 3 ...
-#>  $ satisfaction_3      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 3 4 3 3 4 5 3 4 4 ...
-#>  $ satisfaction_4      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 3 4 4 4 5 4 3 2 4 ...
-#>  $ satisfaction_5      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 7 4 5 5 5 4 3 6 4 6 ...
+#>  $ satisfaction_1      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 4 5 3 4 5 4 5 4 6 ...
+#>  $ satisfaction_2      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 3 4 3 4 6 6 4 4 4 ...
+#>  $ satisfaction_3      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 3 4 4 3 4 5 3 4 4 ...
+#>  $ satisfaction_4      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 7 3 4 4 4 5 4 3 2 4 ...
+#>  $ satisfaction_5      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 7 4 5 4 5 4 3 5 3 6 ...
 ```
 
 The dataset contains 200 respondents, 30 Likert-scale items (10 each for
@@ -490,12 +490,13 @@ survey_clean <- corrPrune(
 cat(sprintf("Reduced from %d → %d variables\n",
             ncol(survey_numeric),
             ncol(survey_clean)))
-#> Reduced from 31 → 4 variables
+#> Reduced from 31 → 7 variables
 
 # Which items were kept?
 selected <- attr(survey_clean, "selected_vars")
 print(selected)
-#> [1] "age"            "satisfaction_9" "engagement_1"   "loyalty_3"
+#> [1] "age"            "satisfaction_1" "engagement_1"   "loyalty_1"     
+#> [5] "loyalty_5"      "loyalty_7"      "loyalty_10"
 ```
 
 The pruning reduced the questionnaire substantially while ensuring age
@@ -520,7 +521,7 @@ cat(sprintf("Satisfaction: %d/10 items kept\n", satisfaction_kept))
 cat(sprintf("Engagement: %d/10 items kept\n", engagement_kept))
 #> Engagement: 1/10 items kept
 cat(sprintf("Loyalty: %d/10 items kept\n", loyalty_kept))
-#> Loyalty: 1/10 items kept
+#> Loyalty: 4/10 items kept
 ```
 
 Good balance: all three constructs retained representation, ensuring the
@@ -595,21 +596,24 @@ summary(model_survey)
 #> 
[…]
