### 2025-12-10 6f72a05
Add badges and remove em-dashes from vignettes
- Add CRAN status and R-CMD-check badges to README (3 total)
- Replace em-dashes with commas/colons in vignettes:
  - workflows.Rmd: "— see" → ", see"
  - quickstart.Rmd: "strict—all" → "strict: all"
  - theory.Rmd: "Key Points —" → "Key Points:" (6 occurrences)
  - theory.Rmd: explanatory em-dashes → commas
- Rebuild pkgdown site

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
