+++ vignettes/advanced.Rmd
@@ -15,7 +15,8 @@ knitr::opts_chunk$set(
   comment = "#>",
   fig.width = 7,
   fig.height = 5,
-  dev = "svg"
+  dev = "svglite",
+  fig.ext = "svg"
 )
 library(corrselect)
 ```
@@ -1033,6 +1034,8 @@ pipeline <- function(data, response) {
 - `?modelPrune` - Model-based pruning
 - `?corrSelect` - Exact subset enumeration
 
+## Session Info
 
-
-
+```{r}
+sessionInfo()
+```
