+++ vignettes/workflows.Rmd
@@ -15,13 +15,18 @@ knitr::opts_chunk$set(
   comment = "#>",
   fig.width = 7,
   fig.height = 5,
+  dev = "svglite",
+  fig.ext = "svg",
   cache = TRUE  # Cache expensive computations to speed up builds
 )
+library(corrselect)
 ```
 
 ## Overview
 
 This vignette demonstrates integration of corrselect into complete modeling workflows, from raw predictors to final models under correlation constraints. Four applied settings illustrate interface selection and workflow composition. See `vignette("quickstart")` for interface descriptions and `vignette("theory")` for mathematical foundations.
+
+**Functions used**: `corrPrune()`, `modelPrune()`, see `vignette("quickstart")` for full signatures and parameter details.
 
 Each workflow showcases different aspects of the package:
 
@@ -48,7 +53,6 @@ This approach balances model fit with interpretability and numerical stability.
 ### Data
 
 ```{r}
-library(corrselect)
 data(bioclim_example)
 
 # Data structure
@@ -59,7 +63,7 @@ head(names(bioclim_example))
 summary(bioclim_example$species_richness)
 ```
 
-The dataset contains 500 sampling locations with 19 bioclimatic predictors and a species richness response. The bioclimatic variables are standard WorldClim metrics (BIO1-BIO19) measuring temperature and precipitation patterns.
+The dataset contains 100 sampling locations with 19 bioclimatic predictors and a species richness response. The bioclimatic variables are standard WorldClim metrics (BIO1-BIO19) measuring temperature and precipitation patterns.
 
 ### Correlation-based pruning
 
@@ -201,9 +205,9 @@ print(comparison)
 
 **Key insights**:
 
-- The full model has severe multicollinearity (κ > 1000)
-- Correlation-based pruning reduces κ by ~10x while losing minimal fit
-- VIF-based refinement achieves excellent stability (κ < 100) with only marginal R² decrease
+- The full model has substantial multicollinearity (high κ)
+- Correlation-based pruning reduces κ while losing minimal fit
+- VIF-based refinement can further improve stability with only marginal R² decrease
 
 ### Visualization
 
@@ -375,7 +379,7 @@ dim(survey_example)
 str(survey_example[, 1:10])  # First 10 columns
 ```
 
-The dataset contains 500 respondents, 30 Likert-scale items (10 each for satisfaction, engagement, loyalty), plus demographics (age, gender, education) and an overall satisfaction outcome.
+The dataset contains 200 respondents, 30 Likert-scale items (10 each for satisfaction, engagement, loyalty), plus demographics (age, gender, education) and an overall satisfaction outcome.
 
 ### Prune with protected variables
 
@@ -409,7 +413,7 @@ selected <- attr(survey_clean, "selected_vars")
 print(selected)
 ```
 
-The pruning reduced the questionnaire from 31 items to ~10 items while ensuring age was retained. The remaining items span all three constructs, avoiding the loss of entire domains.
+The pruning reduced the questionnaire substantially while ensuring age was retained. The remaining items span all three constructs, avoiding the loss of entire domains.
 
 ### Construct coverage
 
@@ -537,25 +541,30 @@ The dataset contains gene expression measurements for 200 genes across 100 sampl
 The greedy algorithm iteratively selects variables, prioritizing those with the lowest maximum correlation to already-selected variables. This heuristic runs in O(p²) time compared to the exponential complexity of exact methods.
 
 ```{r}
+library(microbenchmark)
+
[…]
