### 2025-12-10 efc09af
Remove R-CMD-check badge from README
