### 2025-12-10 ce573dd
Update JOSS paper for v3.0 and gitignore build artifacts
- Rewrite paper.md to cover new v3.0 features:
  - corrPrune() for association-based pruning
  - modelPrune() for VIF-based model pruning
  - Custom engine support (lme4, glmmTMB, INLA, etc.)
  - Fast C++ greedy algorithm
- Add paper/paper.pdf and paper/jats/ to .gitignore
- Remove paper.pdf from tracking (JOSS generates it)

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
