+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "corrselect: Exhaustive variable subset selection based on correlation and association matrices"
+title: "corrselect: Fast and flexible predictor pruning for data analysis and modeling"
 authors:
   - name: Gilles Colling
     orcid: 0000-0003-3070-6066
@@ -15,49 +15,60 @@ bibliography: paper.bib
 
 # Summary
 
-`corrselect` [@cran] is a model-agnostic R package for selecting variable subsets whose pairwise correlations or associations do not exceed a user-defined threshold. Instead of returning a single heuristic solution, it enumerates all maximal admissible subsets. This allows users to select subsets before model fitting, avoiding the common problems of highly correlated or associated predictors, which inflate variance estimates, destabilize coefficient estimates, and obscure the relative importance of variables. The package also supports forced inclusion of user-specified predictors (`forced_in`), ensuring that key variables are retained while admissibility constraints govern the remainder.
+`corrselect` [@cran] is an R package for reducing multicollinearity and redundancy in predictor sets. It provides two complementary approaches: (1) high-level pruning functions that return a single optimal subset, and (2) exhaustive enumeration of all maximal admissible subsets. The package handles both numeric and mixed-type data, supports forced inclusion of key predictors, and integrates with standard R modeling workflows including mixed-effects models.
 
-The package supports both numeric and mixed-type data. Correlation-based workflows include measures such as Pearson, Spearman, Kendall, and biweight midcorrelation [@Langfelder2008], which take values in $[-1,1]$. Association-based workflows use measures normalized to $[0,1]$ for consistent thresholding, including distance correlation [@Szekely2007; @Szekely2009], the maximal information coefficient [@Reshef2011], ANOVA $\eta^2$, and Cramér’s V.
+Version 3.0 introduces `corrPrune()` for association-based pruning and `modelPrune()` for VIF-based model pruning, while retaining the original exhaustive enumeration functions (`corrSelect()`, `assocSelect()`, `MatSelect()`). A fast C++ greedy algorithm enables efficient pruning for large predictor sets (p > 100), while exact graph-theoretic algorithms guarantee complete enumeration when exhaustive search is feasible.
 
 # Statement of Need
 
 Collinearity among predictors is common in applied modeling and can degrade inference and prediction [@Dormann2013]. Popular utilities such as `caret::findCorrelation()` apply greedy, order-dependent filtering and return a single solution. Embedded and wrapper methods like the elastic net [@ZouHastie2005] or recursive feature elimination [@Witten2009] can be powerful but couple selection to a specific model and reduce transparency.
 
-`corrselect` instead formulates a global admissible set problem. Given variables $X_1,\dots,X_p$ and pairwise measures $r_{ij}$, the goal is to find all maximal subsets $S$ such that
+`corrselect` addresses these limitations through two interfaces. For routine workflows, `corrPrune()` and `modelPrune()` provide simple, deterministic pruning with a single function call. For exhaustive exploration, the package formulates a global admissible set problem: given variables $X_1,\dots,X_p$ and pairwise measures $r_{ij}$, find all maximal subsets $S$ such that
 
 $$
 |r_{ij}| \le t \quad \text{for all } i \ne j \in S ,
 $$
 
-with a user threshold $t \in (0,1)$. The software supports mixed variable types, optional forced inclusion of key predictors, and exhaustive coverage of all maximal solutions.
+with a user threshold $t \in (0,1)$. This dual approach balances practicality with methodological rigor.
 
 # Functionality
 
-Three core functions implement the main subset selection tasks:
+## High-Level Pruning Functions
 
-- `corrSelect()` takes a numeric data frame, computes pairwise correlations in $[-1,1]$, and selects admissible subsets at threshold $t$.
[…]
