### 2025-12-10 #9539 @gcol33
Version 3.0.2, which is the latest cran version. I updated the paper.md file accordingly
