### 2025-12-11 a260ae3
Rebuild site without R-CMD-check badge
