+++ docs/articles/advanced.md
@@ -137,14 +137,14 @@ p_values <- c(10, 20, 50, 100, 200, 300, 500, 1000)
 benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
-#> 1   10           0.5            0.3
-#> 2   20           0.6            0.5
-#> 3   50           1.5            0.7
-#> 4  100           4.6            1.4
-#> 5  200          23.6            3.5
-#> 6  300          88.2            7.0
-#> 7  500         319.0           15.7
-#> 8 1000            NA           58.4
+#> 1   10           0.8            0.8
+#> 2   20           1.2            0.7
+#> 3   50           2.2            1.5
+#> 4  100           6.9            2.6
+#> 5  200          29.3            5.6
+#> 6  300         104.1           11.9
+#> 7  500         364.8           25.8
+#> 8 1000            NA          101.3
 ```
 
 ``` r
@@ -740,11 +740,11 @@ time2 <- median(microbenchmark(
 )$time) / 1e6  # Convert nanoseconds to milliseconds
 
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
-#> Recomputing each time: 4.2 ms
+#> Recomputing each time: 7.0 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
-#> Precomputed matrix: 2.0 ms
+#> Precomputed matrix: 2.9 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
-#> Speedup: 2.1x faster
+#> Speedup: 2.4x faster
 ```
 
 **Use precomputed matrices when**:
