### 2025-12-11 d3c1f5f
Add MIT license badge to README
