### 2025-12-16 c2d1216
Add mobile CSS for full-width plots on small screens
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
