+++ docs/articles/advanced.md
@@ -137,14 +137,14 @@ p_values <- c(10, 20, 50, 100, 200, 300, 500, 1000)
 benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
-#> 1   10           0.5            0.3
-#> 2   20           0.7            0.4
-#> 3   50           1.6            0.8
-#> 4  100           4.8            1.4
-#> 5  200          22.7            3.6
-#> 6  300          85.1            6.7
-#> 7  500         312.3           16.0
-#> 8 1000            NA           58.9
+#> 1   10           0.8            0.5
+#> 2   20           1.2            0.9
+#> 3   50           2.3            1.3
+#> 4  100           6.6            2.5
+#> 5  200          29.3            5.9
+#> 6  300         107.1           11.6
+#> 7  500         368.9           27.2
+#> 8 1000            NA          100.5
 ```
 
 ``` r
@@ -740,11 +740,11 @@ time2 <- median(microbenchmark(
 )$time) / 1e6  # Convert nanoseconds to milliseconds
 
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
-#> Recomputing each time: 4.3 ms
+#> Recomputing each time: 7.0 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
-#> Precomputed matrix: 2.2 ms
+#> Precomputed matrix: 2.9 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
-#> Speedup: 1.9x faster
+#> Speedup: 2.4x faster
 ```
 
 **Use precomputed matrices when**:
