+++ docs/reference/survey_example.md
@@ -80,17 +80,6 @@ data(survey_example)
 # This dataset has mixed types: numeric (age, overall_satisfaction),
 # factors (gender, education), and ordered factors (Likert items)
 str(survey_example[, 1:10])
-#> 'data.frame':    200 obs. of  10 variables:
-#>  $ respondent_id       : int  1 2 3 4 5 6 7 8 9 10 ...
-#>  $ age                 : num  38 32 18 18 19 39 33 26 26 42 ...
-#>  $ gender              : Factor w/ 3 levels "Female","Male",..: 2 3 1 2 2 1 1 2 2 1 ...
-#>  $ education           : Ord.factor w/ 4 levels "High School"<..: 3 1 4 2 2 1 1 1 2 3 ...
-#>  $ overall_satisfaction: num  58 40 44 40 58 67 61 49 51 52 ...
-#>  $ satisfaction_1      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 4 5 3 4 5 4 5 4 6 ...
-#>  $ satisfaction_2      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 3 4 3 4 6 6 4 4 4 ...
-#>  $ satisfaction_3      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 6 3 4 4 3 4 5 3 4 4 ...
-#>  $ satisfaction_4      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 7 3 4 4 4 5 4 3 2 4 ...
-#>  $ satisfaction_5      : Ord.factor w/ 7 levels "1"<"2"<"3"<"4"<..: 7 4 5 4 5 4 3 5 3 6 ...
 
 # \donttest{
 # Use assocSelect() for mixed-type data pruning
@@ -99,6 +88,5 @@ pruned <- assocSelect(survey_example[, -1],  # Exclude respondent_id
                       threshold = 0.8,
                       method_ord_ord = "spearman")
 length(attr(pruned, "selected_vars"))
-#> [1] 0
 # }
 ```
