+++ docs/reference/genes_example.md
@@ -91,7 +91,6 @@ data(genes_example)
 gene_data <- genes_example[, -(1:2)]  # Exclude ID and outcome
 pruned <- corrPrune(gene_data, threshold = 0.8, mode = "greedy")
 ncol(pruned)  # Reduced from 200 to ~50 genes
-#> [1] 177
 
 # Use pruned genes for classification
 pruned_with_outcome <- data.frame(
