### 2025-12-16 6c8866a
Use shared build script and rebuild pkgdown site
- Switch to centralized build_pkgdown.R from ~/.R
- Update pkgdown config with svglite figures and proper bg color
- Rebuild site with theme-aware SVG processing

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
