+++ build_site.R
@@ -1,589 +1,1 @@
-# build_site.R - Build pkgdown site with theme-aware SVG post-processing
-#
-# Usage: source("build_site.R") or Rscript build_site.R
-#
-# This script builds the pkgdown site and then post-processes all SVG files
-# to make their backgrounds and text theme-aware (light/dark mode responsive).
-
-# Theme colors matching the site's code blocks
-# Light mode
-LIGHT_BG <- "#F5F6F8"       # code-block background
-LIGHT_BORDER <- "#DFD7CA"   # code-block border
-LIGHT_TEXT <- "#3E3F3A"     # body text color
-
-# Dark mode
-DARK_BG <- "#343739"        # code-block background
-DARK_BORDER <- "#495057"    # code-block border
-DARK_TEXT <- "#DFD7CA"      # body text color
-
-# Sandstone theme font stack
-FONT_FAMILY <- '"Roboto", -apple-system, BlinkMacSystemFont, "Segoe UI", system-ui, sans-serif'
-
-#' Make an SVG file theme-aware
-#'
-#' Handles both base R svg() and svglite output formats.
-#' For svglite: Modifies background rect, CSS rules for strokes/text.
-#' For base R: Finds white-filled backgrounds and black text, adds CSS classes.
-#'
-#' @param svg_path Path to SVG file
-#' @param light_bg Background color for light mode
-#' @param light_border Border color for light mode
-#' @param light_text Text color for light mode
-#' @param font_family Font family to use
-#' @param verbose Print status
-#' @return Invisible logical indicating if changes were made
-make_svg_theme_aware <- function(svg_path,
-                                 light_bg = LIGHT_BG,
-                                 light_border = LIGHT_BORDER,
-                                 light_text = LIGHT_TEXT,
-                                 font_family = FONT_FAMILY,
-                                 verbose = TRUE) {
-  svg_lines <- readLines(svg_path, warn = FALSE)
-  svg_content <- paste(svg_lines, collapse = "\n")
-  original <- svg_content
-
-  # Skip if already processed AND has correct background AND no issues
-  # (check for svg-bg class which indicates background was properly set)
-  has_white_strokes <- grepl("stroke: #FFFFFF;", svg_content, fixed = TRUE)
-  # Check for incorrectly replaced white text (text with fill: #F5F6F8)
-  has_wrong_text_fill <- grepl("<text [^>]*fill: #F5F6F8", svg_content)
-  # Check if rect rule has wrong stroke (black instead of theme color)
-  has_wrong_rect_stroke <- grepl("\\.svglite rect \\{[^}]*stroke: #000000", svg_content)
-  # Check for old CSS variable approach that needs updating
-  has_css_variables <- grepl("var\\(--svg-(stroke|bg)-color", svg_content)
-  # Check for combined rule (rect not split out) that needs updating
-  has_combined_rule <- grepl("\\.svglite line.*\\.svglite rect.*\\.svglite circle \\{", svg_content)
-  # Check for legend boxes with fill: #F5F6F8 but no explicit stroke
-  has_legend_without_stroke <- grepl("<rect [^>]*style='stroke-width:[^']*fill: #F5F6F8;'", svg_content)
-  # Check for legend boxes that still have white fill (need to be themed)
-  has_white_legend_fill <- grepl("<rect [^>]*style='[^']*fill: #FFFFFF;", svg_content)
-  # Check for old stroke-based svg-bg rect (now we use CSS border instead)
-  has_old_svg_bg_stroke <- grepl("class='svg-bg'[^>]*stroke-width:", svg_content)
[…]
