### 2025-12-16 f151e83
Add SVG copy button feature to pkgdown site
- Add copy-to-clipboard button for SVG plots (matches code block button)
- Button appears on hover over plots
- Update extra.css with button styling for light/dark mode

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
