### 2025-12-16 gcol33/corrselect#5 @dansmith01
`README.md` examples have warnings and errors
```
Warning messages:
1: In .compute_vif(model, engine, fixed_effects) :
  Design matrix has no columns after removing intercept. Engine: lme4, Fixed effects: x1, x2
2: In max(violations[removable], na.rm = TRUE) :
  no non-missing arguments to max; returning -Inf
3: In .compute_vif(model, engine, fixed_effects) :
  Design matrix has no columns after removing intercept. Engine: lme4, Fixed effects: x1
4: In modelPrune(y ~ x1 + x2 + (1 | group), data = df, engine = "lme4",  :
  Pruning would remove all predictors. Stopping with current set.

Error in cor(mtcars) : 'x' must be numeric
```

openjournals/joss-reviews#9539
