### 2025-12-16 gcol33/corrselect#6 @dansmith01
Installation instructions
Consider adding `install.packages('corrselect')` as an example installation method to `README.md`, either in place of or in addition to `pak::pak("gcol33/corrselect")` / `remotes::install_github("gcol33/corrselect")`.

Also: `vignettes/quickstart.Rmd` references a non-existant repo in `devtools::install_github("GillesColling/corrselect")`.

openjournals/joss-reviews#9539
