### 2025-12-16 gcol33/corrselect#4 @dansmith01
Currently, `covr::report()` reports code coverage at 80.37%

I'd recommend adding some tests for `src/method_spectral.cpp`, which has 0% coverage.
