### 2025-12-16 6fea7d9
Add GitHub Actions CI workflows and remove orphaned spectral method
- Add R-CMD-check workflow for cross-platform testing (macOS, Windows, Ubuntu)
- Add test-coverage workflow for codecov integration
- Remove unused spectral clustering method (method_spectral.cpp/h)
- Clean up build artifacts from git tracking (Rcheck/, cache files, tar.gz)
- Update .gitignore to prevent future tracking of build artifacts

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
