### 2025-12-16 997542a
Add R-CMD-check and codecov badges to README
