### 2025-12-16 b7c896d
Expand test coverage from 84.5% to 87.4%
- Add 63 new tests across corrPrune, corrSelect, assocSelect, and MatSelect
- Test validation paths, edge cases, and error handling
- Add tests for mixed-type data handling in assocSelect
- Add comprehensive MatSelect validation tests
- Fix incorrect test expectation for low correlation case

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
