+++ tests/testthat/test-assocSelect.R
@@ -100,3 +100,204 @@ test_that("assocSelect can use eta for numeric-factor", {
   res <- assocSelect(df, threshold = 1.0)
   expect_s4_class(res, "CorrCombo")
 })
+
+# ===========================================================================
+# Additional edge case tests for assocSelect
+# ===========================================================================
+
+test_that("assocSelect validates threshold argument", {
+  df <- data.frame(a = rnorm(10), b = rnorm(10))
+
+  expect_error(
+    assocSelect(df, threshold = "0.5"),
+    "threshold.*must be a single numeric"
+  )
+
+  expect_error(
+    assocSelect(df, threshold = c(0.5, 0.7)),
+    "threshold.*must be a single numeric"
+  )
+
+  expect_error(
+    assocSelect(df, threshold = NA),
+    "threshold.*must be a single numeric"
+  )
+
+  expect_error(
+    assocSelect(df, threshold = 0),
+    "threshold.*must be in the range"
+  )
+
+  expect_error(
+    assocSelect(df, threshold = 1.5),
+    "threshold.*must be in the range"
+  )
+})
+
+test_that("assocSelect validates minimum columns", {
+  df <- data.frame(a = rnorm(10))
+
+  expect_error(
+    assocSelect(df, threshold = 0.7),
+    "at least two columns"
+  )
+})
+
+test_that("assocSelect handles character columns (converted to factors)", {
+  set.seed(2001)
+  df <- data.frame(
+    a = rnorm(15),
+    b = sample(c("x", "y", "z"), 15, replace = TRUE),
+    stringsAsFactors = FALSE
+  )
+
+  res <- assocSelect(df, threshold = 0.8)
+  expect_s4_class(res, "CorrCombo")
+})
+
+test_that("assocSelect handles logical columns (converted to factors)", {
+  set.seed(2002)
+  df <- data.frame(
+    a = rnorm(15),
+    b = sample(c(TRUE, FALSE), 15, replace = TRUE)
+  )
+
+  res <- assocSelect(df, threshold = 0.8)
+  expect_s4_class(res, "CorrCombo")
+})
+
+test_that("assocSelect handles integer columns (converted to numeric)", {
+  set.seed(2003)
+  df <- data.frame(
+    a = as.integer(sample(1:100, 15, replace = TRUE)),
+    b = as.integer(sample(1:100, 15, replace = TRUE))
+  )
+
+  res <- assocSelect(df, threshold = 0.8)
+  expect_s4_class(res, "CorrCombo")
+})
+
+test_that("assocSelect handles ordered-ordered pairs", {
+  set.seed(2004)
+  df <- data.frame(
+    a = ordered(sample(c("low", "med", "high"), 15, replace = TRUE)),
+    b = ordered(sample(c("small", "large"), 15, replace = TRUE))
+  )
+
+  res <- assocSelect(df, threshold = 0.9)
+  expect_s4_class(res, "CorrCombo")
+})
+
+test_that("assocSelect handles ordered-factor pairs", {
+  set.seed(2005)
+  df <- data.frame(
+    a = ordered(sample(1:3, 15, replace = TRUE)),
+    b = factor(sample(c("A", "B", "C"), 15, replace = TRUE))
+  )
+
+  res <- assocSelect(df, threshold = 0.9)
+  expect_s4_class(res, "CorrCombo")
+})
+
+test_that("assocSelect handles numeric-ordered pairs", {
+  set.seed(2006)
+  df <- data.frame(
+    a = rnorm(15),
[…]
