+++ tests/testthat/test-corrSelect.R
@@ -79,16 +79,29 @@ test_that("force_in accepts numeric indices", {
   expect_true(all(c("x", "z") %in% unlist(res@subset_list)))
 })
 
-test_that("returns empty when no pair meets threshold", {
-  df <- data.frame(x = rnorm(10), y = -rnorm(10))
-  res <- corrSelect(df, threshold = 0.01, method = "els")
+test_that("returns subsets when correlation is below threshold", {
+  set.seed(999)
+  # Create two nearly uncorrelated variables
+  df <- data.frame(x = rnorm(100), y = rnorm(100))
+  res <- corrSelect(df, threshold = 0.5, method = "els")
 
-  # Should get an empty CorrCombo (no subsets of size ≥ 2)
+  # When correlation is low (below threshold), both variables CAN be in the same subset
+  expect_s4_class(res, "CorrCombo")
+  # Should have at least one subset containing both variables
+  expect_gte(length(res@subset_list), 1)
+})
+
+test_that("returns empty when all pairs exceed threshold", {
+  # Create perfectly correlated variables
+  x <- 1:10
+  y <- x  # Perfect correlation
+  df <- data.frame(x = x, y = y)
+  res <- corrSelect(df, threshold = 0.5, method = "els")
+
+  # With r = 1.0 and threshold = 0.5, no pair meets threshold
+  # So we should get no subsets of size >= 2
   expect_s4_class(res, "CorrCombo")
   expect_length(res@subset_list, 0)
-  expect_equal(res@avg_corr, numeric(0))
-  expect_equal(res@min_corr, numeric(0))
-  expect_equal(res@max_corr, numeric(0))
 })
 
 
@@ -198,4 +211,293 @@ test_that("ELS can get more than one combination (named columns)", {
     found <- any(vapply(subsets, function(x) identical(x, exp), logical(1)))
     expect_true(found, info = paste("ELS missing subset", paste(exp, collapse = ",")))
   }
+})
+
+# ============================================================================
+# Edge case tests (reviewer request)
+# ============================================================================
+
+test_that("identity matrix returns single subset with all variables", {
+
+  # Pure identity matrix: all off-diagonal correlations = 0
+
+  # All variables are uncorrelated, so the only maximal subset is all variables
+  n <- 5
+  identity_mat <- diag(n)
+  colnames(identity_mat) <- paste0("V", 1:n)
+  rownames(identity_mat) <- colnames(identity_mat)
+
+
+  # Test with Bron-Kerbosch
+  res_bk <- MatSelect(identity_mat, threshold = 0.5, method = "bron-kerbosch")
+  expect_equal(length(res_bk@subset_list), 1L)
+  expect_equal(sort(res_bk@subset_list[[1]]), paste0("V", 1:n))
+
+  # Test with ELS
+  res_els <- MatSelect(identity_mat, threshold = 0.5, method = "els")
+  expect_equal(length(res_els@subset_list), 1L)
+  expect_equal(sort(res_els@subset_list[[1]]), paste0("V", 1:n))
+
+  # Both algorithms should agree
+  expect_setequal(res_bk@subset_list[[1]], res_els@subset_list[[1]])
+})
+
+test_that("perfect duplicate (r=1.0) variables are separated into different subsets", {
+  # Two perfectly correlated variables (r = 1.0) should never appear together
+
+  # in the same subset when threshold < 1.0
+  n <- 4
+  cor_mat <- diag(n)
+  # V1 and V2 are perfectly correlated
[…]

[…]
