### 2025-12-16 a08d9f7
v3.0.4: Improve test coverage to 93.44%
- Remove dead C++ code (isValidAddition, isValidCombination) from utils.cpp/utils.h
- Add edge case tests for ELS algorithm (force_in validation, threshold boundaries)
- Add edge case tests for association methods (Cramer's V sparse tables, eta)
- Add tests for corrPrune lexicographic tiebreaker and factor handling
- Add tests for modelPrune custom engine error handling and VIF edge cases
- Coverage improved from 91.86% to 93.44%

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
