+++ paper/paper.md
@@ -21,7 +21,7 @@ Version 3.0 introduces `corrPrune()` for association-based pruning and `modelPru
 
 # Statement of Need
 
-Collinearity among predictors is common in applied modeling and can degrade inference and prediction [@Dormann2013]. Popular utilities such as `caret::findCorrelation()` apply greedy, order-dependent filtering and return a single solution. Embedded and wrapper methods like the elastic net [@ZouHastie2005] or recursive feature elimination [@Witten2009] can be powerful but couple selection to a specific model and reduce transparency.
+Collinearity among predictors is common in applied modeling and can degrade inference and prediction [@Dormann2013]. Popular utilities such as `caret::findCorrelation()` apply greedy, order-dependent filtering and return a single solution—typically removing the variable with the highest mean correlation at each step. This heuristic approach discards potentially useful subsets and provides no guarantee of optimality. Correlation-based filter methods such as FCBF [@YuLiu2003] similarly employ greedy strategies. Embedded and wrapper methods like the elastic net [@ZouHastie2005] or recursive feature elimination [@Witten2009] can be powerful but couple selection to a specific model and reduce transparency.
 
 `corrselect` addresses these limitations through two interfaces. For routine workflows, `corrPrune()` and `modelPrune()` provide simple, deterministic pruning with a single function call. For exhaustive exploration, the package formulates a global admissible set problem: given variables $X_1,\dots,X_p$ and pairwise measures $r_{ij}$, find all maximal subsets $S$ such that
 
@@ -29,7 +29,7 @@ $$
 |r_{ij}| \le t \quad \text{for all } i \ne j \in S ,
 $$
 
-with a user threshold $t \in (0,1)$. This dual approach balances practicality with methodological rigor.
+with a user threshold $t \in (0,1)$. Unlike greedy methods that return a single result, `corrselect` in exact mode enumerates *all* maximal admissible subsets, enabling researchers to explore the full solution space. This dual approach balances practicality with methodological rigor.
 
 # Functionality
 
@@ -51,15 +51,36 @@ Three functions implement complete subset enumeration:
 
 All enumeration functions return a `CorrCombo` object containing maximal subsets, summary statistics, and standard methods (`print`, `summary`, `as.data.frame`). The helper function `corrSubset()` extracts filtered data frames from results.
 
+A typical workflow demonstrates the key difference from greedy methods:
+
+```r
+library(corrselect)
+result <- corrSelect(mtcars, threshold = 0.7)
+result
+#> CorrCombo object with 12 maximal subsets
+#>   Threshold: 0.7 | Correlation method: pearson
+#>   Sizes: 6, 5, 5, 5, ... | Avg |r|: 0.30, 0.27, 0.30, 0.31, ...
+
+as.data.frame(result)[1:3, c("subset", "size", "avg_corr")]
+#>                          subset size avg_corr
+#> 1 mpg, cyl, drat, qsec, vs, am    6     0.30
+#> 2     mpg, drat, qsec, vs, gear    5     0.27
+#> 3      mpg, hp, drat, qsec, am    5     0.30
+```
+
+Unlike `caret::findCorrelation()` which returns a single variable set, `corrSelect()` reveals all 12 equally valid solutions, enabling informed selection based on domain knowledge.
+
 ## Algorithms
 
-The package implements three algorithms in C++:
+The admissible set problem is mathematically equivalent to finding all maximal cliques in the "compatibility graph" where edges connect variable pairs with $|r_{ij}| \le t$, or equivalently, all maximal independent sets in the "conflict graph" where edges connect pairs exceeding the threshold. This connection to well-studied graph problems provides a solid theoretical foundation.
+
[…]
