+++ paper/paper.bib
@@ -2,7 +2,7 @@
   title        = {corrselect: Correlation-Based Variable Subset Selection},
   author       = {Gilles Colling},
   year         = {2025},
-  note         = {R package version 2.0.1},
+  note         = {R package version 3.0.3},
   url          = {https://CRAN.R-project.org/package=corrselect},
   doi          = {10.32614/CRAN.package.corrselect}
 }
@@ -60,7 +60,7 @@
   number  = {6062},
   pages   = {1518--1524},
   doi     = {10.1126/science.1205438}
-}}
+}
 
 @article{Langfelder2008,
   title   = {WGCNA: An R package for weighted correlation network analysis},
@@ -104,4 +104,24 @@
   publisher = {Springer},
   year      = {2010},
   doi       = {10.1007/978-3-642-17517-6_36}
+}
+
+@article{igraph,
+  title   = {The igraph software package for complex network research},
+  author  = {Csardi, G{\'a}bor and Nepusz, Tam{\'a}s},
+  journal = {InterJournal, Complex Systems},
+  volume  = {1695},
+  number  = {5},
+  pages   = {1--9},
+  year    = {2006},
+  url     = {https://igraph.org}
+}
+
+@inproceedings{YuLiu2003,
+  author    = {Yu, Lei and Liu, Huan},
+  title     = {Feature Selection for High-Dimensional Data: A Fast Correlation-Based Filter Solution},
+  booktitle = {Proceedings of the 20th International Conference on Machine Learning (ICML-03)},
+  pages     = {856--863},
+  year      = {2003},
+  publisher = {AAAI Press}
 }
