+++ tests/testthat/test-modelPrune.R
@@ -640,3 +640,445 @@ test_that("modelPrune custom engine can implement non-VIF diagnostics", {
   expect_equal(length(attr(result, "selected_vars")), 3)
   expect_equal(attr(result, "engine"), "pvalue_pruner")
 })
+
+# ===========================================================================
+# Additional coverage tests for modelPrune.R
+# ===========================================================================
+
+test_that("modelPrune custom engine with wrong length diagnostics", {
+  df <- mtcars
+
+  bad_engine <- list(
+    name = "wrong_length",
+    fit = function(formula, data, ...) stats::lm(formula, data = data),
+    diagnostics = function(model, fixed_effects) {
+      # Return named vector with wrong length
+      c(cyl = 1)  # Only one value when expecting 2
+    }
+  )
+
+  expect_error(
+    modelPrune(mpg ~ cyl + disp, data = df, engine = bad_engine, limit = 5),
+    "missing names for"
+  )
+})
+
+test_that("modelPrune custom engine criterion parameter ignored with message", {
+  df <- mtcars
+
+  custom_engine <- list(
+    name = "test",
+    fit = function(formula, data, ...) stats::lm(formula, data = data),
+    diagnostics = function(model, fixed_effects) {
+      setNames(rep(1, length(fixed_effects)), fixed_effects)
+    }
+  )
+
+  # Setting criterion to non-vif should produce message
+  expect_message(
+    modelPrune(mpg ~ cyl, data = df, engine = custom_engine,
+               criterion = "aic", limit = 10),
+    "criterion.*ignored"
+  )
+})
+
+test_that("modelPrune handles max_steps warning", {
+  set.seed(1201)
+  df <- data.frame(
+    y = rnorm(100),
+    x1 = rnorm(100),
+    x2 = rnorm(100),
+    x3 = rnorm(100)
+  )
+  # Make x2 and x3 correlated
+  df$x3 <- df$x2 * 0.95 + rnorm(100, sd = 0.1)
+
+  # With max_steps = 1 and low limit, should warn
+  expect_warning(
+    modelPrune(y ~ x1 + x2 + x3, data = df, limit = 2, max_steps = 1),
+    "max_steps"
+  )
+})
+
+test_that("modelPrune handles max_steps NA error", {
+  df <- mtcars
+
+  expect_error(
+    modelPrune(mpg ~ cyl, data = df, max_steps = NA),
+    "'max_steps' must be"
+  )
+})
+
+test_that("modelPrune handles multiple predictor vectors with force_in", {
+  df <- mtcars
+
+  # Multiple force_in variables that satisfy threshold
+  result <- modelPrune(mpg ~ cyl + disp + hp + wt,
+                       data = df, force_in = c("cyl", "wt"), limit = 50)
+
+  expect_true(all(c("cyl", "wt") %in% attr(result, "selected_vars")))
+})
+
+test_that("modelPrune handles limit = 0 error", {
+  df <- mtcars
+
+  expect_error(
+    modelPrune(mpg ~ cyl, data = df, limit = 0),
+    "'limit' must be positive"
+  )
+})
+
+test_that("modelPrune with interaction terms", {
+  df <- mtcars
+
+  result <- modelPrune(mpg ~ cyl * disp, data = df, limit = 100)
+
+  expect_s3_class(result, "data.frame")
+  expect_true("mpg" %in% names(result))
+})
+
+test_that("modelPrune glm with poisson family", {
+  set.seed(1202)
+  df <- data.frame(
[…]

[…]
