### 2025-12-16 gcol33/corrselect#4 @gcol33
Thank you for reviewing the test coverage. The file src/method_spectral.cpp was removed since it was a legacy functionality and I decided to keep the package identity as enumerating all possible subsets instead of using heuristics, the current test coverage is approximately 93.44%.
