### 2025-12-16 d6e289b
v3.0.5: Fix CRAN check issues
- Add svglite and data.table to Suggests (required by vignettes and tests)
- Fix codecov URL redirect in README

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
