### 2025-12-17 976f1b0
Fix mobile CSS for svg-container wrapper
The SVG copy button feature wraps SVGs in .svg-container divs,
which broke the mobile full-width CSS selectors. Added selectors
for p:has(> .svg-container) and p > .svg-container > svg.inline-svg
to restore rectangular borders on mobile.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
