+++ docs/articles/advanced.md
@@ -137,14 +137,14 @@ p_values <- c(10, 20, 50, 100, 200, 300, 500, 1000)
 benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
-#> 1   10           0.5            0.2
-#> 2   20           0.6            0.4
-#> 3   50           1.6            0.7
-#> 4  100           4.4            1.4
-#> 5  200          22.4            3.4
-#> 6  300          86.6            6.8
-#> 7  500         312.8           15.7
-#> 8 1000            NA           57.7
+#> 1   10           0.8            0.4
+#> 2   20           0.9            0.6
+#> 3   50           1.6            1.3
+#> 4  100           3.8            2.9
+#> 5  200           9.1            5.8
+#> 6  300          20.6            9.1
+#> 7  500          51.8           21.7
+#> 8 1000            NA           63.8
 ```
 
 ``` r
@@ -740,11 +740,11 @@ time2 <- median(microbenchmark(
 )$time) / 1e6  # Convert nanoseconds to milliseconds
 
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
-#> Recomputing each time: 4.5 ms
+#> Recomputing each time: 5.3 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
-#> Precomputed matrix: 2.1 ms
+#> Precomputed matrix: 1.3 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
-#> Speedup: 2.2x faster
+#> Speedup: 4.2x faster
 ```
 
 **Use precomputed matrices when**:
@@ -1363,7 +1363,7 @@ sessionInfo()
 #> [1] stats     graphics  grDevices utils     datasets  methods   base     
 #> 
 #> other attached packages:
-#> [1] microbenchmark_1.5.0 corrselect_3.0.2    
+#> [1] microbenchmark_1.5.0 corrselect_3.0.5    
 #> 
 #> loaded via a namespace (and not attached):
 #>  [1] digest_0.6.37     desc_1.4.3        R6_2.6.1          fastmap_1.2.0    
