### 2025-12-17 c8d37ed
Fix mobile CSS: target .svg-container directly
Browser auto-closes <p> before <div>, so .svg-container is a sibling
not a child. Target .svg-container directly for full-width mobile.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
