### 2025-12-17 f5a4960
Fix invalid HTML generation for SVG containers
- Updated inline_svgs_in_html() to detect when <img> is inside <p> tag
- Closes </p> before <div> and reopens <p> after for valid HTML
- Regenerated docs with fix from ~/.R/build_pkgdown.R

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
