### 2025-12-17 728cb3a
Fix mobile CSS selectors for svg-container
Regenerated extra.css with correct mobile CSS selectors that
target .svg-container directly (not as child of p).

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
