+++ docs/articles/advanced.md
@@ -138,13 +138,13 @@ benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
 #> 1   10           0.8            0.4
-#> 2   20           1.1            0.7
-#> 3   50           1.9            1.5
+#> 2   20           1.0            0.6
+#> 3   50           1.7            1.2
 #> 4  100           3.5            2.3
-#> 5  200           8.5            5.5
-#> 6  300          19.2            9.2
-#> 7  500          49.0           20.1
-#> 8 1000            NA           66.4
+#> 5  200           9.2            5.4
+#> 6  300          19.5            9.4
+#> 7  500          52.8           20.4
+#> 8 1000            NA           65.8
 ```
 
 ``` r
@@ -740,11 +740,11 @@ time2 <- median(microbenchmark(
 )$time) / 1e6  # Convert nanoseconds to milliseconds
 
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
-#> Recomputing each time: 5.2 ms
+#> Recomputing each time: 5.3 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
-#> Precomputed matrix: 1.3 ms
+#> Precomputed matrix: 1.2 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
-#> Speedup: 4.0x faster
+#> Speedup: 4.3x faster
 ```
 
 **Use precomputed matrices when**:
