### 2025-12-17 f4fe757
Skip energy package on oldrel-1 in CI
The gsl package (dependency of energy) now requires R >= 4.5.0,
causing CI failures on R 4.4.x. Skip energy installation on oldrel-1
since it's only a Suggests dependency and tests already guard with
skip_if_not_installed().

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
