+++ .github/workflows/R-CMD-check.yml
@@ -42,6 +42,7 @@ jobs:
         with:
           extra-packages: any::rcmdcheck
           needs: check
+          pak-ignore: ${{ matrix.config.r == 'oldrel-1' && 'energy' || '' }}
 
       - uses: r-lib/actions/check-r-package@v2
         with:
