### 2025-12-17 5809f29
Fix README examples to avoid modifying mtcars
- Use local copy (mtcars_glm) for GLM example instead of modifying mtcars
- Add defensive numeric filtering for cor(mtcars) example

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
