### 2025-12-17 gcol33/corrselect#5 @gcol33
There was a bug in the README examples. The GLM example modified mtcars directly by adding a factor column (am_binary), which caused the later cor(mtcars) call to fail when running the examples sequentially. I've fixed this by using a local copy (mtcars_glm) for the GLM example and adding defensive filtering in the correlation matrix example.
