### 2025-12-17 b9034a2
Fix energy dependency skip on oldrel-1
Use sed to remove energy from DESCRIPTION before dependency resolution
on R 4.4.x, since pak-ignore parameter doesn't exist.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
