### 2025-12-17 f38071c
Address JOSS reviewer feedback on paper.md
- Strengthen caret comparison: clarify practical difference (one subset vs many)
- Add graph theory context in Statement of Need (maximal cliques)
- Add benchmark data: p=100 (<1s), p=150 (~19s), p=175 (~3min), p=200 (~17min)
- Reorganize Functionality section (model-agnostic vs model-based)
- Clarify FCBF addresses a different problem (supervised filter)
- Fix typo in .Rbuildignore (remove duplicate calude entry)

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
