### 2025-12-17 ac31402
Refine JOSS paper structure and content
- Reorganize Functionality into Model-Agnostic and Model-Based sections
- Clarify distinction from supervised methods (FCBF)
- Document custom greedy algorithm vs caret::findCorrelation()
- Add benchmark timings for exact enumeration scaling
- Fix Related Work to properly scope model-agnostic claims

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
