+++ paper/paper.md
@@ -73,7 +73,7 @@ For exact enumeration, the package implements two algorithms natively in C++ (no
 - **Bron-Kerbosch**: The classical maximal clique enumeration algorithm [@Bron1973], used by default for unrestricted enumeration.
 - **Eppstein-Löffler-Strash (ELS)**: A near-optimal algorithm for sparse graphs [@Eppstein2010], used when `force_in` seeds are specified.
 
-Both exact methods ensure non-redundant and complete enumeration of admissible subsets. However, maximal clique enumeration is NP-hard in the general case, and exact mode may become impractically slow on large or densely connected problems. On an Intel i9-14900K with correlations clustered near the threshold, $p = 100$ completed in under 1 second, $p = 150$ in ~19 seconds, $p = 175$ in ~3 minutes, and $p = 200$ in ~17 minutes. Exact mode is therefore recommended only for moderate-sized problems ($p \le 100$).
+Both exact methods ensure non-redundant and complete enumeration of admissible subsets. However, maximal clique enumeration is NP-hard in the general case, and exact mode may become impractically slow on large or densely connected problems. On a modern desktop (Intel i9-14900K) with correlations clustered near the threshold, $p = 100$ completed in under 1 second, $p = 150$ in ~19 seconds, $p = 175$ in ~3 minutes, and $p = 200$ in ~17 minutes. Exact mode is therefore recommended only for moderate-sized problems ($p \le 100$).
 
 For larger predictor sets or low thresholds where exact enumeration becomes infeasible, `corrPrune(mode = "greedy")` provides a fast polynomial-time alternative. Unlike `caret::findCorrelation()` which removes the variable with the highest mean correlation, this custom greedy algorithm iteratively removes the variable involved in the most threshold violations, with ties broken by maximum and then average association. This runs in $O(p^2 \times k)$ time where $k$ is the number of variables removed.
 
