### 2025-12-17 49a843d
Remove AI-style phrasing from paper.md
- Rewrite FCBF sentence for clarity
- Replace "balances practicality with methodological rigor" with simpler phrasing
- Remove generic "provides a solid theoretical foundation" filler

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
