### 2025-12-17 9c183ed
Fix remaining AI-style phrasing in paper.md
- "In contrast to" -> "Unlike"
- "extends this space" -> "goes further"
- Remove tricolon in Related Work section

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
