### 2025-12-17 ac58734
Polish JOSS paper wording
- Simplify example intro text
- Minor phrasing improvements throughout

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
