### 2025-12-17 22de794
Fix R CMD check energy dependency warnings on oldrel-1
- Use energy=?ignore to skip installing energy on R 4.4.x
- Set _R_CHECK_FORCE_SUGGESTS_=false for oldrel-1 checks
- Keep energy in DESCRIPTION so R CMD check doesn't flag ::calls

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
