### 2025-12-17 c090556
Fix flaky test and energy dependency handling
- Add set.seed(42) to assocSelect force_in test for reproducibility
- Use energy=?ignore on oldrel-1 instead of removing from DESCRIPTION
- Set _R_CHECK_FORCE_SUGGESTS_=false for oldrel-1 checks

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
