### 2025-12-19 gcol33/corrselect#7 @danStich
Package-level documentation
It would be nice to have some package-level documentation that is easy to browse. This was one of the first commands I tried to run just to get an overview of the package and see what functions and objects were included. I often do this first time using an R package. The package index page looks great on the packagedown site, and to be honest the package structure is not so complex that it is strictly necessary. I just happened to be offline when I started working through the R package for this review and couldn't find package-level documentation beyond auto-complete suggestions from typing `corrselect::`, (e.g., no documentation for corrselect using help('corrselect').

[openjournals/joss-reviews/issues/9539](https://github.com/openjournals/joss-reviews/issues/9539)
