### 2025-12-19 gcol33/corrselect#8 @danStich
Association-based example in README
Suggest changing comment from "Check what was removed" to "Check what was retained". This confused me at first as I was following along with the README examples to get started, but it is clear from documentation for the function that the selection returned is for those variables that were retained.

[openjournals/joss-reviews/issues/9539](https://github.com/openjournals/joss-reviews/issues/9539)
