### 2025-12-19 gcol33/corrselect#9 @danStich
Deterministic tie-breaking (order of entry)
Is there a way to make the deterministic tie-breaking method clearer to the user up front and/or to provide some suggestions for dealing with this if only through manual rearrangement? It is easy to miss that it is based on order of entry. It might be helpful to just provide the user with a message saying that they should consider relevance of the tied predictors or re-examine candidate set using, for example, the 'exact' method or another stochastic approach that does not rely on order of entry. The Details section of the help file was the only place I saw this mentioned. I am not sure if such a message already exists but didn't run into it with my test data sets. This has been an issue historically in my field, probably others, usually associated with uninformed use of automated tools by (non-statistical) subject-matter experts.

[openjournals/joss-reviews/issues/9539](https://github.com/openjournals/joss-reviews/issues/9539)
