### 2025-12-20 7166cb4
v3.0.6: Address reviewer feedback (issues #7-10)
- Add package-level documentation (?corrselect now works) (#7)
- Fix README comment: "removed" -> "retained" (#8)
- Document tie-breaking behavior in corrPrune() (#9)

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
