### 2025-12-20 b36d038
Fix R-CMD-check macOS dependency installation failures
Add cache-version: 2 to setup-r-dependencies steps to force cache refresh
and avoid corrupted binary package downloads for e1071 and reformulas.

Co-authored-by: gcol33 <25009600+gcol33@users.noreply.github.com>
