+++ .github/workflows/R-CMD-check.yml
@@ -44,12 +44,14 @@ jobs:
         with:
           extra-packages: any::rcmdcheck
           needs: check
+          cache-version: 2
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         if: matrix.config.r == 'oldrel-1'
         with:
           extra-packages: any::rcmdcheck, energy=?ignore
           needs: check
+          cache-version: 2
 
       - uses: r-lib/actions/check-r-package@v2
         if: matrix.config.r != 'oldrel-1'
