### 2025-12-20 cb17655
Add explicit pak-version and R_REMOTES_NO_ERRORS_FROM_WARNINGS to workflow
Co-authored-by: gcol33 <25009600+gcol33@users.noreply.github.com>
