### 2025-12-20 8d294b9
Apply same improvements to test-coverage workflow
Co-authored-by: gcol33 <25009600+gcol33@users.noreply.github.com>
