+++ .github/workflows/test-coverage.yml
@@ -12,6 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
 
     steps:
       - uses: actions/checkout@v4
@@ -19,11 +20,24 @@ jobs:
       - uses: r-lib/actions/setup-r@v2
         with:
           use-public-rspm: true
+      
+      # Clean up any stale pak cache that might cause issues
+      - name: Clean pak cache
+        shell: Rscript {0}
+        run: |
+          if (requireNamespace("pak", quietly = TRUE)) {
+            try(pak::cache_clean(), silent = TRUE)
+          }
+        continue-on-error: true
 
-      - uses: r-lib/actions/setup-r-dependencies@v2
+      - name: Install dependencies
+        uses: r-lib/actions/setup-r-dependencies@v2
         with:
           extra-packages: any::covr
           needs: coverage
+          pak-version: stable
+        env:
+          PKG_SYSREQS: true
 
       - name: Test coverage
         run: |
