### 2025-12-20 c43e374
Fix R-CMD-check macOS failure by installing GSL
The energy package requires GSL (GNU Scientific Library) which is not
automatically installed on macOS by r-lib/actions (only Linux has
automatic sysreqs handling). Add explicit brew install gsl step.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
