+++ .github/workflows/R-CMD-check.yml
@@ -38,6 +38,11 @@ jobs:
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
 
+      # Install GSL for energy package on macOS (not auto-installed by r-lib/actions)
+      - name: Install macOS system dependencies
+        if: runner.os == 'macOS'
+        run: brew install gsl
+
       # On oldrel-1, skip energy (gsl requires R >= 4.5.0)
       - uses: r-lib/actions/setup-r-dependencies@v2
         if: matrix.config.r != 'oldrel-1'
