### 2025-12-20 1ccabd6
Skip Bioconductor packages on macOS to fix R-CMD-check
WGCNA, GO.db, preprocessCore, and impute have complex system
dependencies that fail to install on macOS GitHub Actions runners.
Skip these optional Suggests on macOS and use _R_CHECK_FORCE_SUGGESTS_=false.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
