### 2025-12-20 564f890
Use hard dependencies only on macOS R-CMD-check
Instead of ignoring specific Bioconductor packages, only install hard
dependencies (Imports/Depends) on macOS. This avoids complex dependency
resolution issues with WGCNA ecosystem and other Suggests.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
