### 2025-12-20 56ce704
Pin macOS version to macos-14 instead of macos-latest
macos-latest may have been updated with breaking changes. Using
a specific version for stability.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
