### 2025-12-20 ec10c80
Try macos-13 (Intel) instead of macos-14 (ARM)
macos-14 (ARM) may have cache or architecture issues. Testing with
macos-13 (Intel) for stability.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
