### 2025-12-20 4e3965e
Invalidate corrupted R package cache with cache-version: 2
The macOS runner failure was caused by corrupted binary package cache
for e1071 and reformulas packages. Bumping cache-version forces fresh
package downloads.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
