### 2025-12-21 973e41f
Add concurrency settings to prevent job cancellation
Set cancel-in-progress: false to prevent automatic cancellation of
running workflow jobs when new commits are pushed.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
