+++ .github/workflows/R-CMD-check.yml
@@ -7,6 +7,10 @@ on:
     branches: [main, master]
   workflow_dispatch:
 
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: false
+
 jobs:
   R-CMD-check:
     runs-on: ${{ matrix.config.os }}
