### 2025-12-21 dbe1fd3
Switch from deprecated macos-13 to macos-14
macOS-13 runner was fully deprecated on December 4, 2025. Using
macos-14 (ARM64) which is the current stable macOS runner.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
