### 2025-12-21 a90ecb4
Fix macOS CI: use =?ignore syntax instead of invalid 'hard' dependencies
