### 2025-12-21 c1b46db
Fix macOS CI: use pak-ignore parameter for Bioconductor packages
