+++ .github/workflows/R-CMD-check.yml
@@ -59,7 +59,8 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         if: runner.os == 'macOS'
         with:
-          extra-packages: any::rcmdcheck, GO.db=?ignore, WGCNA=?ignore, preprocessCore=?ignore, impute=?ignore
+          extra-packages: any::rcmdcheck
+          pak-ignore: GO.db, WGCNA, preprocessCore, impute
           cache-version: 2
           needs: check
 
