### 2025-12-21 a17c840
Fix macOS CI: use multiline =?ignore syntax and bump cache
