+++ .github/workflows/R-CMD-check.yml
@@ -59,9 +59,13 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         if: runner.os == 'macOS'
         with:
-          extra-packages: any::rcmdcheck
-          pak-ignore: GO.db, WGCNA, preprocessCore, impute
-          cache-version: 2
+          extra-packages: |
+            any::rcmdcheck
+            GO.db=?ignore
+            WGCNA=?ignore
+            preprocessCore=?ignore
+            impute=?ignore
+          cache-version: 3
           needs: check
 
       # oldrel-1: skip energy (gsl requires R >= 4.5.0)
