### 2025-12-21 2021537
Fix macOS CI: use sed to remove Bioconductor/energy deps from DESCRIPTION
