### 2025-12-21 32d6b1f
Fix macOS CI: disable RSPM to avoid missing ARM64 binaries
