+++ .github/workflows/R-CMD-check.yml
@@ -37,10 +37,18 @@ jobs:
       - uses: r-lib/actions/setup-pandoc@v2
 
       - uses: r-lib/actions/setup-r@v2
+        if: runner.os != 'macOS'
         with:
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
+
+      # macOS: don't use RSPM (missing ARM64 binaries cause download failures)
+      - uses: r-lib/actions/setup-r@v2
+        if: runner.os == 'macOS'
+        with:
+          r-version: ${{ matrix.config.r }}
+          use-public-rspm: false
 
       # macOS: remove Bioconductor packages and energy from DESCRIPTION before dependency resolution
       - name: Remove problematic deps on macOS
