### 2025-12-21 4816aa4
Fix macOS CI: delete lockfile to force fresh dependency resolution
