### 2025-12-21 a2e74a7
Fix macOS CI: remove caret/Boruta and force source installs
