### 2025-12-21 f43d42a
Fix macOS CI: set PKGTYPE=source to force source installs
