+++ .github/workflows/R-CMD-check.yml
@@ -50,11 +50,11 @@ jobs:
           r-version: ${{ matrix.config.r }}
           use-public-rspm: false
 
-      # macOS: remove problematic packages from DESCRIPTION
+      # macOS: remove packages with missing ARM64 binaries from DESCRIPTION
       - name: Remove problematic deps on macOS
         if: runner.os == 'macOS'
         run: |
-          sed -i.bak '/GO.db/d; /WGCNA/d; /preprocessCore/d; /impute/d; /energy/d; /caret/d; /Boruta/d' DESCRIPTION
+          sed -i.bak '/GO.db/d; /WGCNA/d; /preprocessCore/d; /impute/d; /energy/d; /caret/d; /Boruta/d; /glmnet/d' DESCRIPTION
 
       # Standard install (Linux, Windows - not macOS, not oldrel-1)
       - uses: r-lib/actions/setup-r-dependencies@v2
@@ -64,14 +64,14 @@ jobs:
           cache-version: 2
           needs: check
 
-      # macOS: force source installs to avoid missing binary issues
+      # macOS: force source installs only (RSPM lacks ARM64 binaries for some packages)
       - uses: r-lib/actions/setup-r-dependencies@v2
         if: runner.os == 'macOS'
         env:
-          R_COMPILE_AND_INSTALL_PACKAGES: always
+          PKGTYPE: source
         with:
           extra-packages: any::rcmdcheck
-          cache-version: 3
+          cache-version: 5
           needs: check
 
       # oldrel-1: skip energy (gsl requires R >= 4.5.0)
