### 2025-12-21 9ac9b5b
Fix macOS CI: use remotes::install_deps with type=source
