### 2025-12-21 3e7edca
Fix macOS CI: use remotes with CRAN binaries (not source)
