### 2025-12-21 c67c90a
Fix macOS CI: add CRAN mirror setting
