### 2025-12-21 b6adcdf
Skip tibble tests if tibble not installed
Fixes macOS CI failure where tibble's dependency pkgconfig was
missing, causing test errors instead of skips.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
