+++ .github/workflows/R-CMD-check.yml
@@ -94,6 +94,7 @@ jobs:
         with:
           upload-snapshots: true
           build_args: 'c("--no-manual","--compact-vignettes=gs+qpdf")'
+          error-on: '"error"'
 
       - uses: r-lib/actions/check-r-package@v2
         if: matrix.config.r == 'oldrel-1'
