### 2025-12-21 f82c77f
Fix macOS CI: install Rcpp explicitly before deps
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
