### 2025-12-21 41fa642
Fix macOS CI: install vignette deps explicitly
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
