+++ .github/workflows/R-CMD-check.yml
@@ -69,8 +69,8 @@ jobs:
         if: runner.os == 'macOS'
         run: |
           options(repos = c(CRAN = "https://cloud.r-project.org"))
-          install.packages(c("remotes", "Rcpp", "rcmdcheck"))
-          remotes::install_deps(dependencies = TRUE, upgrade = "never")
+          install.packages(c("remotes", "Rcpp", "rcmdcheck", "knitr", "rmarkdown", "svglite"))
+          remotes::install_deps(dependencies = TRUE)
         shell: Rscript {0}
 
       # oldrel-1: skip energy (gsl requires R >= 4.5.0)
