### 2025-12-21 774bfc7
macOS CI: keep CRAN packages, only remove Bioconductor deps
WGCNA, energy, caret, Boruta, and glmnet now have ARM64 binaries
on CRAN. Only Bioconductor packages (GO.db, preprocessCore, impute)
need to be removed as they lack ARM64 binaries.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
