+++ tests/testthat/test-modelPrune.R
@@ -1082,3 +1082,346 @@ test_that("modelPrune custom engine diagnostics returns non-numeric", {
     "must return a numeric vector"
   )
 })
+
+# ===========================================================================
+# Additional edge case tests for full coverage
+# ===========================================================================
+
+test_that("modelPrune handles Inf VIF values (near-perfect collinearity)", {
+  set.seed(9101)
+  n <- 100
+  x1 <- rnorm(n)
+  # Create near-perfect collinearity
+  df <- data.frame(
+    y = rnorm(n),
+    x1 = x1,
+    x2 = x1 + rnorm(n, sd = 1e-10),  # Essentially identical
+    x3 = rnorm(n)
+  )
+
+  # Should handle Inf VIF gracefully
+  result <- suppressWarnings(
+    modelPrune(y ~ x1 + x2 + x3, data = df, limit = 5)
+  )
+
+  expect_s3_class(result, "data.frame")
+})
+
+test_that("modelPrune handles VIF with all near-collinear predictors", {
+  set.seed(9102)
+  n <- 100
+  x1 <- rnorm(n)
+  df <- data.frame(
+    y = rnorm(n),
+    x1 = x1,
+    x2 = x1 + rnorm(n, sd = 1e-8),
+    x3 = x1 + rnorm(n, sd = 1e-8)
+  )
+
+  # All have Inf VIF essentially
+  result <- suppressWarnings(
+    modelPrune(y ~ x1 + x2 + x3, data = df, limit = 10)
+  )
+
+  expect_s3_class(result, "data.frame")
+})
+
+test_that("modelPrune tie-breaking removes last variable in formula order", {
+  set.seed(9103)
+  n <- 100
+  x1 <- rnorm(n)
+  # Create two equally collinear pairs
+  df <- data.frame(
+    y = rnorm(n),
+    a = x1,
+    b = x1 + rnorm(n, sd = 0.1),
+    c = rnorm(n),
+    d = rnorm(n)
+  )
+
+  result <- modelPrune(y ~ a + b + c + d, data = df, limit = 5)
+
+  # b should be removed (same VIF as a, but later in formula)
+  expect_true("a" %in% attr(result, "selected_vars") ||
+              "b" %in% attr(result, "selected_vars"))
+})
+
+test_that("modelPrune errors when no fixed effects remain", {
+  df <- mtcars
+
+  # Create custom engine that always returns high diagnostics
+  always_high_engine <- list(
+    name = "always_high",
+    fit = function(formula, data, ...) stats::lm(formula, data = data),
+    diagnostics = function(model, fixed_effects) {
+      setNames(rep(100, length(fixed_effects)), fixed_effects)
+    }
+  )
+
+  # This should warn about removing all predictors
+  expect_warning(
+    modelPrune(mpg ~ cyl, data = df, engine = always_high_engine, limit = 1),
+    "would remove all"
+  )
+})
+
+test_that("modelPrune handles design matrix with factor predictors", {
+  set.seed(9104)
+  n <- 60
+  df <- data.frame(
+    y = rnorm(n),
+    x1 = rnorm(n),
+    cat = factor(sample(c("A", "B", "C"), n, replace = TRUE)),
+    x2 = rnorm(n)
+  )
+
+  result <- modelPrune(y ~ x1 + cat + x2, data = df, limit = 10)
+
+  expect_s3_class(result, "data.frame")
+})
+
+test_that("modelPrune VIF computation with missing column match", {
+  set.seed(9105)
+  n <- 50
[…]
