### 2025-12-22 0b74afe
Add grouped pruning, new measures, and condition_number criterion
corrPrune:
- Add `by` parameter for grouped pruning (compute associations per group)
- Add `group_q` parameter to control quantile aggregation across groups
- Add bicor, distance, maximal measures for numeric data (require
  WGCNA, energy, minerva packages respectively)

modelPrune:
- Add condition_number criterion using SVD-based computation
- Update documentation to reflect supported criteria

Tests:
- Add tests for grouped pruning functionality and edge cases
- Add tests for condition_number criterion
- Update validation test
[…]
