### 2025-12-22 4410396
Bump version to 3.0.7 and update documentation
- Update NEWS.md with 3.0.7 release notes
- Update quickstart vignette with new parameters (by, group_q, criterion)
- Add grouped pruning section to advanced vignette
- Add condition_number documentation to advanced vignette
- Rebuild pkgdown website

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
