+++ docs/reference/corrPrune.md
@@ -133,9 +133,23 @@ pairwise associations below `threshold` in *all* groups. With
 groups.
 
 **Mode Selection**: Exact mode guarantees finding all maximal subsets
-and returns the largest one (with deterministic tie-breaking). Greedy
-mode is faster but approximate, using a deterministic removal strategy
-based on association scores.
+and returns the largest one. Greedy mode is faster but approximate,
+using an iterative removal strategy based on association scores.
+
+**Tie-Breaking**: When multiple subsets or variables are equally good,
+deterministic tie-breaking is applied:
+
+- **Exact mode**: Selects by (1) largest subset size, (2) lowest average
+  correlation, (3) alphabetically first variable names. Column order
+  does not affect the result.
+
+- **Greedy mode**: Removes the variable with (1) most constraint
+  violations, (2) highest max association, (3) highest average
+  association, (4) lowest column index. Column order can influence the
+  result when earlier criteria are tied.
+
+To see all maximal subsets instead of a single selection, use
+[`corrSelect()`](https://gillescolling.com/corrselect/reference/corrSelect.md).
 
 ## See also
 
