+++ man/modelPrune.Rd
@@ -40,9 +40,16 @@ of diagnostic scores (one per fixed effect, higher values = worse)
 }}
 
 \item{criterion}{Character string specifying the diagnostic criterion for pruning.
-For built-in engines, only \code{"vif"} (Variance Inflation Factor) is supported.
+For built-in engines, supported values are:
+\itemize{
+\item \code{"vif"} (default): Variance Inflation Factor. Measures how much the variance
+of a coefficient is inflated due to collinearity. Values > 5-10 indicate
+problematic multicollinearity.
+\item \code{"condition_number"}: Condition indices based on singular value decomposition
+of the design matrix. Higher values indicate greater collinearity.
 For custom engines, this parameter is ignored (diagnostics are computed by the
-engine's \code{diagnostics} function). Default: \code{"vif"}.}
+engine's \code{diagnostics} function).
+}}
 
 \item{limit}{Numeric scalar. Maximum allowed value for the criterion.
 Predictors with diagnostic values exceeding this limit are iteratively removed.
