### 2025-12-22 eeeca02
Improve test coverage and fix mixed-type bugs
Bug fixes:
- Fix numeric-numeric pairs in mixed-type data (use pearson, not auto measure)
- Fix numeric-ordered pairs (convert ordered to numeric before correlation)

New tests:
- Add tests for bicor, distance, maximal measures (with skip_if_not)
- Add tests for mixed-type data: numeric-ordered, ordered-ordered, factor-factor
- Add tests for lme4 and glmmTMB engines
- Add tests for condition_number with edge cases

Coverage improved from 92.14% to 93.38%

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.
[…]
