### 2025-12-22 e91b8c2
Add more edge case tests for improved coverage
- Add tests for single-level factors and constant variables in assocSelect
- Add tests for invalid force_in validation in findAllMaxSets
- Add tests for lexicographic tie-breaking scenarios
- Add tests for perfectly collinear data and quasi-separation in GLMs

Coverage: 93.38% -> 93.52%
findAllMaxSets.R: now at 100% coverage

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
