### 2025-12-22 1700777
Add synthetic data tests for edge cases
- Add lexicographic tie-breaking test (hit lines 502-508 in corrPrune)
- Add tests for sparse contingency tables
- Add more eta edge case tests for assocSelect
- Add corrSelect tests for bicor/distance/maximal

Coverage: 93.52% -> 93.96%
corrPrune.R: 95.17% -> 97.40%

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
