### 2025-12-22 66412a4
Release v3.1.0: Bug fixes and test coverage improvements
Bug fixes:
- corrPrune: Fixed numeric-numeric pair handling in mixed-type data
- corrPrune: Fixed numeric-ordered pair conversion for Spearman correlation

Test coverage (92% → 94.10%):
- Added tests for optional measures (bicor, distance, maximal)
- Added tests for lme4/glmmTMB engines
- Added chi-squared and VIF edge case tests
- Added lexicographic tie-breaking tests
- Added mixed-type data tests
- Added synthetic edge case tests for assocSelect/corrPrune/modelPrune
- findAllMaxSets.R at 100%, corrPrune.R at 98%
- 731 tests passing

🤖
[…]
