+++ tests/testthat/test-modelPrune.R
@@ -1706,3 +1706,250 @@ test_that("modelPrune GLM handles quasi-separation", {
   )
   expect_s3_class(result, "data.frame")
 })
+
+
+# ===========================================================================
+# Chi-squared NA edge case (line 350 in corrPrune)
+# ===========================================================================
+
+test_that("corrPrune handles chi2 NA from empty cells in contingency table", {
+  set.seed(11001)
+  n <- 20
+
+  # Create factors that produce a contingency table where chi2 might be NA
+  # This happens when expected frequencies are very low
+  fac1 <- factor(c(rep("A", 18), "B", "C"))
+  fac2 <- factor(c(rep("X", 18), "Y", "Z"))
+
+  df <- data.frame(fac1 = fac1, fac2 = fac2, num1 = rnorm(n))
+
+  # Should handle gracefully even with problematic contingency table
+  result <- corrPrune(df, threshold = 0.99)
+  expect_s3_class(result, "data.frame")
+})
+
+# ===========================================================================
+# VIF: Column matching with factors (lines 622, 627-633)
+# ===========================================================================
+
+test_that("modelPrune VIF handles factor with many levels", {
+  set.seed(11002)
+  n <- 100
+
+  df <- data.frame(
+    y = rnorm(n),
+    x1 = rnorm(n),
+    x2 = rnorm(n),
+    cat = factor(sample(LETTERS[1:5], n, replace = TRUE))
+  )
+
+  # Factor creates multiple columns in design matrix
+  result <- modelPrune(y ~ x1 + x2 + cat, data = df, limit = 20)
+  expect_s3_class(result, "data.frame")
+})
+
+# ===========================================================================
+# VIF: R-squared is NA (line 669)
+# This happens with perfectly collinear predictors
+# ===========================================================================
+
+test_that("modelPrune handles R-squared NA from collinearity", {
+  set.seed(11003)
+  n <- 50
+
+  x1 <- rnorm(n)
+  df <- data.frame(
+    y = rnorm(n),
+    x1 = x1,
+    x2 = x1 * 2,      # Perfect linear relationship
+    x3 = x1 * 3,      # Perfect linear relationship
+    x4 = rnorm(n)
+  )
+
+  # This creates a situation where R² might be NA or very close to 1
+  expect_warning(
+    result <- modelPrune(y ~ x1 + x2 + x3 + x4, data = df, limit = 5)
+  )
+  expect_s3_class(result, "data.frame")
+})
+
+# ===========================================================================
+# VIF: Single predictor remaining after removal (lines 642-643)
+# ===========================================================================
+
+test_that("modelPrune VIF returns 1 for single predictor", {
+  set.seed(11004)
+  n <- 50
+
+  x1 <- rnorm(n)
+  df <- data.frame(
+    y = rnorm(n),
+    x1 = x1,
+    x2 = x1 + rnorm(n, sd = 0.001)  # Extremely collinear
+  )
+
+  # With very strict limit, should reduce to single predictor
+  result <- suppressWarnings(
+    modelPrune(y ~ x1 + x2, data = df, limit = 1.1)
+  )
+  expect_s3_class(result, "data.frame")
+  # Should have only 1 predictor (VIF = 1 for single predictor)
+  expect_true(length(attr(result, "selected_vars")) >= 1)
+})
+
+# ===========================================================================
+# VIF: Error in lm computation (lines 678-679)
[…]
