### 2026-01-08 5b5b49d
Fix CRAN NOTE: import quantile from stats
- Add quantile to @importFrom stats in corrPrune.R
- Update NAMESPACE accordingly
- Add cookies.txt to .Rbuildignore

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
