+++ NAMESPACE
@@ -18,6 +18,7 @@ importFrom(stats,complete.cases)
 importFrom(stats,cor)
 importFrom(stats,lm)
 importFrom(stats,model.matrix)
+importFrom(stats,quantile)
 importFrom(stats,sd)
 importFrom(stats,setNames)
 importFrom(stats,terms)
