### 2026-01-09 3c8a40c
Remove unpublished JOSS paper section from README
🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
