### 2026-01-10 ed0cf9c
Add CRAN DOI and URL to citation
