### 2026-01-11 d198cff
Add Documentation section to README
Link to pkgdown articles at gillescolling.com/corrselect

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
