### 2026-01-14 b2c749a
Rebuild site with legend text dark mode fix
