+++ docs/articles/quickstart.md
@@ -453,11 +453,11 @@ sessionInfo()
 #> [1] corrselect_3.1.0
 #> 
 #> loaded via a namespace (and not attached):
-#>  [1] digest_0.6.37     desc_1.4.3        R6_2.6.1          fastmap_1.2.0    
-#>  [5] xfun_0.53         cachem_1.1.0      knitr_1.50        htmltools_0.5.8.1
-#>  [9] rmarkdown_2.30    lifecycle_1.0.4   cli_3.6.5         svglite_2.2.2    
-#> [13] sass_0.4.10       pkgdown_2.2.0     textshaping_1.0.3 jquerylib_0.1.4  
+#>  [1] digest_0.6.39     desc_1.4.3        R6_2.6.1          fastmap_1.2.0    
+#>  [5] xfun_0.55         cachem_1.1.0      knitr_1.51        htmltools_0.5.9  
+#>  [9] rmarkdown_2.30    lifecycle_1.0.5   cli_3.6.5         svglite_2.2.2    
+#> [13] sass_0.4.10       pkgdown_2.2.0     textshaping_1.0.4 jquerylib_0.1.4  
 #> [17] systemfonts_1.3.1 compiler_4.5.2    tools_4.5.2       bslib_0.9.0      
-#> [21] evaluate_1.0.5    Rcpp_1.1.0        yaml_2.3.10       jsonlite_2.0.0   
-#> [25] rlang_1.1.6       fs_1.6.6          htmlwidgets_1.6.4
+#> [21] evaluate_1.0.5    Rcpp_1.1.1        yaml_2.3.12       otel_0.2.0       
+#> [25] jsonlite_2.0.0    rlang_1.1.6       fs_1.6.6          htmlwidgets_1.6.4
 ```
