### 2026-01-14 80bc3d5
Preserve colored strokes in dark mode
