### 2026-01-14 7df9f27
Keep dark outlines on histogram bars in dark mode
