+++ docs/articles/advanced.md
@@ -137,14 +137,14 @@ p_values <- c(10, 20, 50, 100, 200, 300, 500, 1000)
 benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
-#> 1   10           0.5            0.3
-#> 2   20           0.7            0.4
-#> 3   50           1.4            0.7
-#> 4  100           4.4            1.5
-#> 5  200          20.8            3.5
-#> 6  300          79.0            6.9
-#> 7  500         275.8           15.8
-#> 8 1000            NA           59.1
+#> 1   10           0.5            0.4
+#> 2   20           0.8            0.4
+#> 3   50           1.5            0.7
+#> 4  100           4.1            1.5
+#> 5  200          20.3            3.2
+#> 6  300          75.3            6.7
+#> 7  500         273.1           15.7
+#> 8 1000            NA           56.1
 ```
 
 ``` r
@@ -800,9 +800,9 @@ time2 <- median(microbenchmark(
 )$time) / 1e6  # Convert nanoseconds to milliseconds
 
 cat(sprintf("Recomputing each time: %.1f ms\n", time1))
-#> Recomputing each time: 4.4 ms
+#> Recomputing each time: 4.0 ms
 cat(sprintf("Precomputed matrix: %.1f ms\n", time2))
-#> Precomputed matrix: 2.3 ms
+#> Precomputed matrix: 2.1 ms
 cat(sprintf("Speedup: %.1fx faster\n", time1 / time2))
 #> Speedup: 1.9x faster
 ```
