### 2026-01-14 e3a082b
Add MathJax line breaking for mobile-friendly equations
