### 2026-01-14 f011683
Fix SVG copy button visibility on mobile (tap-to-show)
