### 2026-01-19 #9539 @Nikoleta-v3
Happy New Year, everyone! @danStich, did you get a chance to look at @gcol33’s responses to your issues?
