### 2026-01-19 #9539 @danStich
Oh yes, the looked great! I am so sorry, I didn't realize I was holding this up.

Dan

----
Daniel S. Stich (he/him)
Professor, Biology Department and Biological Field Station Researcher
Biology Department Graduate Programs Coordinator
113A Perna Science, SUNY Oneonta, NY 13820
Office: 607-436-3734
Cell: 518-860-4107
Email: ***@***.***
Website: danstich.github.io/stich<https://danstich.github.io/stich/index.html>
