### 2026-01-26 #9539 @Nikoleta-v3
@gcol33, thank you for your patience. I went over the paper one last time and opened a pull request to fix an issue in a reference, please see: https://github.com/gcol33/corrselect/pull/13

Also, could you please:

- Let me know whether I should go ahead and use v3.0.2 as the version, or if you would prefer to create and use a new release.
- Archive the reviewed software on Zenodo or a similar service (e.g., Figshare or an institutional repository).
  - Ensure that the archival deposit (e.g., on Zenodo) includes the correct metadata, including:
  - The correct title (matching the paper title)
  - An accurate author list
  - Optional: ORCIDs for each author

Finally, please share the DOI of the archived version here.
