### 2026-01-30 1a68aa4
Update archive_doi to v3.1.0 Zenodo deposit
DOI: 10.5281/zenodo.18429041

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>
