### 2026-01-30 #9539 @gcol33
@Nikoleta-v3,

Thank you for the review and for merging [PR #13.](https://github.com/gcol33/corrselect/pull/13)

Please use v3.1.0 as the authoritative version. This release includes minor fixes made during the review process (including [PR #13.](https://github.com/gcol33/corrselect/pull/13)).

  - Title: corrselect: Fast and flexible predictor pruning for data analysis and modeling
  - Author: Gilles Colling (ORCID: 0000-0003-3070-6066)
  - Version: 3.1.0

DOI: https://doi.org/10.5281/zenodo.18429041
