### 2026-01-31 #9539 @Nikoleta-v3
@gcol33 could you please update the title of the Zenodo archive to match that of the paper: `corrselect: Exhaustive variable subset selection based on correlation and association matrices`? Also, could you add your ORCID to the Zenodo archive?
