### 2026-01-31 #9539 @gcol33
> [@gcol33](https://github.com/gcol33) could you please update the title of the Zenodo archive to match that of the paper: `corrselect: Exhaustive variable subset selection based on correlation and association matrices`? Also, could you add your ORCID to the Zenodo archive?

@Nikoleta-v3 I added the ORCID to the Zenodo archive, I was surprised it wasnt added automatically since they asked me to connect my accounts. Oh well, anyway it's now updated as well as the title under: https://zenodo.org/records/18429041
