### 2026-02-01 #9539 @Nikoleta-v3
Ha! I also have it connected to my account, and I believe it does add it automatically, but maybe only now that I’ve done it manually once. Hopefully in the future it will just work, hahaha.

Thank you again, @gcol33, and sorry to ask one more thing. could you also fix your affiliation to match the paper? Currently it just says “University of Vienna,” but it doesn’t include the department as it appears in the paper. Could you change that manually? Sorry!
