### 2026-02-02 #9539 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
