### 2026-02-24 9bc5136
Fix SVG rendering bug in workflows vignette (figure inside heading)
The Coefficient Comparison figure was rendered inside an <h3> tag due to
a knitr artifact (### prefix on figure output in markdown). Browsers
displayed the SVG CSS and text content as raw visible text instead of
rendering the graphic.

Fix: added defensive guard in build_pkgdown.R that detects and strips
heading wrappers around SVG containers. Rebuilt pkgdown site with cleared
vignette cache.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
