+++ docs/articles/comparison.md
@@ -7,8 +7,11 @@ established methods for multicollinearity management and variable
 selection:
 
 1.  **caret::findCorrelation()** - Greedy correlation-based pruning
+
 2.  **Boruta** - Random forest permutation importance
+
 3.  **glmnet** - L1/L2 regularization (LASSO/Ridge)
+
 4.  **Manual VIF removal** - Iterative variance inflation factor
     thresholding
 
@@ -87,7 +90,9 @@ Block structure present: correlations range from -0.15 to 0.97.
 caret’s `findCorrelation()` applies greedy iterative removal:
 
 1.  Identify pair with maximum $`|r_{ij}|`$
+
 2.  Remove variable with larger mean absolute correlation
+
 3.  Repeat until all $`|r_{ij}| < \tau`$
 
 Non-deterministic: results depend on internal ordering. Typically
@@ -224,9 +229,12 @@ forced variable constraints, mixed data types.
 Boruta tests variable importance via random forest permutation:
 
 1.  Create shadow features (permuted copies)
+
 2.  Fit random forest on original + shadow features
+
 3.  Test:
     $`\text{importance}(X_i) > \max(\text{importance}(\text{shadow}))`$
+
 4.  Iteratively confirm/reject until convergence
 
 **Orthogonal objective**: Boruta selects predictive variables
@@ -329,7 +337,9 @@ glmnet minimizes regularized loss:
 ```
 
 - $`\alpha = 1`$: LASSO (L1 penalty, sparse $`\beta`$)
+
 - $`\alpha = 0`$: Ridge (L2 penalty, shrinkage)
+
 - $`\lambda`$: Cross-validation selected
 
 **Difference**: glmnet performs soft selection (shrinkage) optimizing
@@ -450,7 +460,9 @@ where $`R^2_j`$ results from regressing $`X_j`$ on remaining predictors.
 Thresholds:
 
 - VIF \< 5: Low collinearity
+
 - VIF \< 10: Moderate (acceptable)
+
 - VIF ≥ 10: High (problematic)
 
 Manual approach: iteratively remove max(VIF) until all VIF \< threshold.
@@ -608,11 +620,15 @@ reproducible documentation.
 
 1.  **Maximal clique enumeration**: Optimal retention under
     $`|r_{ij}| < \tau`$ constraint
+
 2.  **Deterministic**: ELS and Bron-Kerbosch algorithms guarantee
     reproducibility
+
 3.  **Flexible**: Forced variables, mixed types, greedy/exact modes
+
 4.  **Unbiased estimates**: Hard removal preserves coefficient
     interpretability
+
 5.  **Model-agnostic**: Correlation-based preprocessing
 
 ### Integrated Workflow
@@ -662,10 +678,13 @@ final_model <- lm(response ~ ., data = model_data[, c("response", important_vars
 
 - [`vignette("quickstart")`](https://gillescolling.com/corrselect/articles/quickstart.md) -
   Interface overview and usage examples
+
 - [`vignette("workflows")`](https://gillescolling.com/corrselect/articles/workflows.md) -
   Domain-specific workflows (genomics, ecology, surveys)
+
 - [`vignette("advanced")`](https://gillescolling.com/corrselect/articles/advanced.md) -
   Algorithm selection and performance tuning
+
 - [`vignette("theory")`](https://gillescolling.com/corrselect/articles/theory.md) -
   Graph-theoretic foundations and formal proofs
 
@@ -700,9 +719,9 @@ sessionInfo()
[…]
