+++ docs/articles/advanced.md
@@ -7,10 +7,14 @@ method developers:
 
 1.  **Understanding the Algorithms** - Exact vs greedy, complexity
     analysis
+
 2.  **Custom Engines** - Integrate any modeling package (INLA, mgcv,
     brms)
+
 3.  **Exact Subset Enumeration** - Multiple maximal subsets
+
 4.  **Performance Optimization** - Speed and memory considerations
+
 5.  **Troubleshooting** - Common issues and solutions
 
 **Target audience**: Users comfortable with R programming and
@@ -46,8 +50,11 @@ cat("Exact mode kept:", ncol(exact_result), "variables\n")
 **Use exact mode when**:
 
 - p ≤ 20 (feasible runtime)
+
 - You need guaranteed optimal solution
+
 - Reproducibility is critical
+
 - You’re writing a paper (justify optimality)
 
 #### Greedy Mode (Heuristic)
@@ -66,8 +73,11 @@ cat("Greedy mode kept:", ncol(greedy_result), "variables\n")
 **Use greedy mode when**:
 
 - p \> 20 (exact becomes slow)
+
 - Speed is priority
+
 - Near-optimal is acceptable
+
 - High-dimensional data (p \> 100)
 
 #### Auto Mode (Recommended)
@@ -137,14 +147,14 @@ p_values <- c(10, 20, 50, 100, 200, 300, 500, 1000)
 benchmark <- benchmark_corrPrune(p_values)
 print(benchmark)
 #>      p exact_time_ms greedy_time_ms
-#> 1   10           0.5            0.3
-#> 2   20           0.6            0.4
-#> 3   50           1.4            0.7
-#> 4  100           4.1            1.3
-#> 5  200          20.1            3.3
-#> 6  300          73.8            6.5
-#> 7  500         266.9           15.5
-#> 8 1000            NA           57.0
+#> 1   10           0.6            0.3
+#> 2   20           0.7            0.4
+#> 3   50           1.6            0.7
+#> 4  100           4.4            1.6
+#> 5  200          21.3            3.5
+#> 6  300          76.7            7.0
+#> 7  500         276.9           16.6
+#> 8 1000            NA           61.7
 ```
 
 ``` r
@@ -237,8 +247,9 @@ cat("Identical:", identical(names(result1), names(result2)), "\n")
 ```
 
 **Tie-breaking rules**: 1. Prefer variables with lower **mean absolute
-correlation** with others 2. If still tied, prefer **lexicographically
-first** (alphabetical)
+correlation** with others
+
+2.  If still tied, prefer **lexicographically first** (alphabetical)
 
 This ensures reproducibility across runs, machines, and R versions.
 
@@ -280,8 +291,10 @@ cat("Selected:", attr(result, "selected_vars"), "\n")
 #### When to Use
 
 - **Multi-site studies**: Correlations may differ across locations
+
 - **Experimental conditions**: Treatment groups may have different
   correlation structures
+
 - **Longitudinal data**: Correlations may change over time periods
 
 ------------------------------------------------------------------------
@@ -298,8 +311,11 @@ not just base R’s [`lm()`](https://rdrr.io/r/stats/lm.html) or `lme4`.
 This enables diagnostic-based pruning (VIF or condition number) for:
 
 - **Bayesian models** (INLA, brms, Stan)
+
 - **Additive models** (mgcv GAMs)
+
 - **Survival models** (coxph, flexsurv)
+
 - **Custom metrics** (AIC, BIC, posterior uncertainty)
 
 #### Built-in Criteria
@@ -309,6 +325,7 @@ available:
 
[…]
