+++ docs/authors.md
@@ -2,7 +2,8 @@
 
 ## Authors
 
-- **[Gilles Colling](https://github.com/gcol33)**. Author, maintainer.
+- **[Gilles Colling](https://github.com/gcol33)**. Author, maintainer,
+  copyright holder. [](https://orcid.org/0000-0003-3070-6066)
 
 ## Citation
 
