### 2026-02-25 #9539 @crvernon
@gcol33 - I just need you to address the following before I accept for publication:

In your archive:

- [ ] The license you have specified in your Zenodo archive must match the license you specify in your GitHub repository.  Please edit the metadata in your Zenodo archive to reflect this.  No new release needed.
- [ ] The title of your archive must match the title of your paper exactly.  Please fix this in the metadata of your archive as well.

Let me know when these changes have been made and after I confirm I will accept.
