### 2026-02-25 #9539 @gcol33
Hi @crvernon , thanks for checking!

Both issues have been addressed in the Zenodo archive metadata:

  - License: Updated to MIT.
  - Title: Updated to "corrselect: Fast and flexible predictor pruning for data analysis and modeling", matching the paper title exactly.

Let me know if anything else is needed.
